Combobox: Move combobox to react
