+++ packages/react-internal/src/components/Autofill/Autofill.test.tsx
@@ -174,6 +174,8 @@ describe('Autofill', () => {
       target: autofill.inputElement!,
       nativeEvent: {
         isComposing: true,
+
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any,
     });
 
@@ -183,6 +185,7 @@ describe('Autofill', () => {
       target: autofill.inputElement!,
       nativeEvent: {
         isComposing: true,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any,
     });
     jest.runOnlyPendingTimers();
@@ -192,6 +195,7 @@ describe('Autofill', () => {
       which: KeyCodes.m,
       nativeEvent: {
         isComposing: true,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any,
     });
     autofill.inputElement!.value = '🆘m';
@@ -199,6 +203,7 @@ describe('Autofill', () => {
       target: autofill.inputElement!,
       nativeEvent: {
         isComposing: true,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any,
     });
 
@@ -208,6 +213,7 @@ describe('Autofill', () => {
       target: autofill.inputElement!,
       nativeEvent: {
         isComposing: false,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any,
     });
     jest.runOnlyPendingTimers();
