--- packages/react-next/src/components/Autofill/index.ts
@@ -1,2 +1,0 @@
-export * from './Autofill';
-export * from './Autofill.types';
