--- packages/react-next/src/components/ComboBox/index.ts
@@ -1,3 +1,0 @@
-export * from './ComboBox';
-export * from './ComboBox.types';
-export * from './VirtualizedComboBox';
