+++ packages/react/src/components/ComboBox/ComboBox.styles.ts
@@ -13,7 +13,7 @@ import {
 } from '../../Styling';
 import { IComboBoxOptionStyles, IComboBoxStyles } from './ComboBox.types';
 
-import { IButtonStyles } from '../../Button';
+import { IButtonStyles } from '../../compat/Button';
 import { memoizeFunction } from '../../Utilities';
 
 const ComboBoxHeight = 32;
