+++ packages/react/src/components/ComboBox/ComboBox.test.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import * as ReactTestUtils from 'react-dom/test-utils';
-import { mount, ReactWrapper } from 'enzyme';
 import * as renderer from 'react-test-renderer';
 import { KeyCodes } from '../../Utilities';
-
-import { ComboBox, IComboBoxState } from './ComboBox';
-import { IComboBox, IComboBoxOption, IComboBoxProps } from './ComboBox.types';
+import { ComboBox } from './ComboBox';
+import { IComboBox, IComboBoxOption } from './ComboBox.types';
 import { SelectableOptionMenuItemType } from '../../SelectableOption';
-import { expectOne, expectMissing, renderIntoDocument } from '@fluentui/react-internal/lib/common/testUtilities';
+import { isConformant } from '../../common/isConformant';
+import { safeCreate } from '@uifabric/test-utilities';
 
 const DEFAULT_OPTIONS: IComboBoxOption[] = [
   { key: '1', text: '1' },
@@ -37,11 +36,6 @@ const RUSSIAN_OPTIONS: IComboBoxOption[] = [
 
 const returnUndefined = () => undefined;
 
-type InputElementWrapper = ReactWrapper<React.InputHTMLAttributes<any>, any>;
-
-let wrapper: ReactWrapper<IComboBoxProps, IComboBoxState, ComboBox> | undefined;
-let domNode: HTMLElement | undefined;
-
 const createNodeMock = (el: React.ReactElement<{}>) => {
   return {
     __events__: {},
@@ -49,20 +43,10 @@ const createNodeMock = (el: React.ReactElement<{}>) => {
 };
 
 describe('ComboBox', () => {
-  afterEach(() => {
-    if (wrapper) {
-      wrapper.unmount();
-      wrapper = undefined;
-    }
-    if (domNode) {
-      try {
-        ReactDOM.unmountComponentAtNode(domNode.parentElement!);
-        domNode.parentElement!.removeChild(domNode);
-      } catch (ex) {
-        // ignore
-      }
-      domNode = undefined;
-    }
+  beforeEach(() => {
+    spyOn(ReactDOM, 'createPortal').and.callFake(element => {
+      return element;
+    });
   });
 
   it('Renders correctly', () => {
@@ -83,38 +67,63 @@ describe('ComboBox', () => {
     expect(tree).toMatchSnapshot();
   });
 
+  it(`renders`, () => {
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} />, wrapper => {
+      expect(wrapper.root).toBeDefined();
+    });
+  });
+
+  isConformant({
+    Component: ComboBox,
+    displayName: 'ComboBox',
+    // Disabled due to being required to mount by safeCreate. A test called "renders" was added here as a replacement.
+    disabledTests: ['component-renders'],
+  });
+
   it('Can flip between enabled and disabled.', () => {
-    wrapper = mount(<ComboBox disabled={false} options={DEFAULT_OPTIONS} />);
+    safeCreate(<ComboBox disabled={false} options={DEFAULT_OPTIONS} />, container => {
+      expect(container.root.findAll(node => node.props.className?.split?.(' ')?.includes?.('is-disabled')).length).toBe(
+        0,
+      );
 
-    expectMissing(wrapper, '.ms-ComboBox.is-disabled');
-    expectOne(wrapper, '[data-is-interactable=true]');
+      expect(
+        container.root.findAll(node => typeof node.type === 'string' && node.props['data-is-interactable'] === true)
+          .length,
+      ).toBe(1);
 
-    wrapper.setProps({ disabled: true });
+      renderer.act(() => {
+        container.update(<ComboBox disabled={true} options={DEFAULT_OPTIONS} />);
+      });
 
-    expectOne(wrapper, '.ms-ComboBox.is-disabled');
-    expectOne(wrapper, '[data-is-interactable=false]');
+      expect(container.root.findAll(node => node.props.className?.split?.(' ')?.includes?.('is-disabled')).length).toBe(
+        2,
+      );
+      expect(
+        container.root.findAll(node => typeof node.type === 'string' && node.props['data-is-interactable'] === false)
+          .length,
+      ).toBe(1);
+    });
   });
 
   it('Renders no selected item in default case', () => {
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} />);
-
-    expect(wrapper.find('input[role="combobox"]').text()).toEqual('');
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('');
+    });
   });
 
   it('Renders a selected item in uncontrolled case', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS} />);
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement: InputElementWrapper = comboBoxRoot.find('input');
-
-    expect(inputElement.props().value).toEqual('1');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('1');
+    });
   });
 
   it('Renders a selected item in controlled case', () => {
-    wrapper = mount(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} />);
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement: InputElementWrapper = comboBoxRoot.find('input');
-
-    expect(inputElement.props().value).toEqual('1');
+    safeCreate(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('1');
+    });
   });
 
   it('Renders a selected item with zero key', () => {
@@ -122,10 +131,10 @@ describe('ComboBox', () => {
       { key: 0, text: 'zero' },
       { key: 1, text: 'one' },
     ];
-    wrapper = mount(<ComboBox selectedKey={0} options={options} />);
-
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    expect(inputElement.props().value).toEqual('zero');
+    safeCreate(<ComboBox selectedKey={0} options={options} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('zero');
+    });
   });
 
   it('changes to a selected key change the input', () => {
@@ -133,13 +142,14 @@ describe('ComboBox', () => {
       { key: 0, text: 'zero' },
       { key: 1, text: 'one' },
     ];
-    wrapper = mount(<ComboBox selectedKey={0} options={options} />);
-
-    expect(wrapper.find('input').props().value).toEqual('zero');
-
-    wrapper.setProps({ selectedKey: 1 });
-
-    expect(wrapper.find('input').props().value).toEqual('one');
+    safeCreate(<ComboBox selectedKey={0} options={options} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('zero');
+      renderer.act(() => {
+        container.update(<ComboBox selectedKey={1} options={options} />);
+      });
+      expect(input.props.value).toEqual('one');
+    });
   });
 
   it('changes to a selected item on key change', () => {
@@ -147,196 +157,242 @@ describe('ComboBox', () => {
       { key: 0, text: 'zero' },
       { key: 1, text: 'one' },
     ];
-    wrapper = mount(<ComboBox selectedKey={0} options={options} />);
-
-    expect(wrapper.find('input').props().value).toEqual('zero');
-
-    wrapper.setProps({ selectedKey: null });
-
-    expect(wrapper.find('input').props().value).toEqual('');
+    safeCreate(<ComboBox selectedKey={0} options={options} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('zero');
+      renderer.act(() => {
+        container.update(<ComboBox selectedKey={null} options={options} />);
+      });
+      expect(input.props.value).toEqual('');
+    });
   });
 
   it('Renders a placeholder', () => {
     const placeholder = 'Select an option';
-    wrapper = mount(<ComboBox placeholder={placeholder} options={DEFAULT_OPTIONS} />);
-
-    const inputElement = wrapper.find('.ms-ComboBox input').getDOMNode() as HTMLInputElement;
-    expect(inputElement.placeholder).toEqual(placeholder);
-    expect(inputElement.value).toEqual('');
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} placeholder={placeholder} />, container => {
+      const inputElement = container.root.findByType('input');
+      expect(inputElement.props.placeholder).toEqual(placeholder);
+      expect(inputElement.props.value).toEqual('');
+    });
   });
 
   it('Does not automatically add new options when allowFreeform is on in controlled case', () => {
-    const componentRef = React.createRef<any>();
-    wrapper = mount(
-      <ComboBox
-        options={DEFAULT_OPTIONS}
-        allowFreeform={true}
-        onChange={returnUndefined}
-        componentRef={componentRef}
-      />,
-    );
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} allowFreeform={true} onChange={returnUndefined} />, container => {
+      const inputElement = container.root.findByType('input');
 
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    inputElement.simulate('input', { target: { value: 'f' } });
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    expect((componentRef.current as ComboBox).state.currentOptions.length).toEqual(DEFAULT_OPTIONS.length);
+      simulateInputEvent(inputElement, 'f');
+      simulateKeydown(inputElement, KeyCodes.enter);
+
+      // open combobox
+      const buttonElement = container.root.findByType('button');
+      ReactTestUtils.act(() => {
+        buttonElement?.props?.onClick();
+      });
+
+      const options = container.root.findAll(
+        node =>
+          node.props.className &&
+          node.props.className.split(' ').includes('ms-ComboBox-option') &&
+          typeof node.type === 'string',
+      );
+
+      expect(options.length).toBe(DEFAULT_OPTIONS.length);
+    });
   });
 
   it('Automatically adds new options when allowFreeform is on in uncontrolled case', () => {
-    const componentRef = React.createRef<any>();
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} allowFreeform={true} componentRef={componentRef} />);
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} allowFreeform={true} />, container => {
+      const inputElement = container.root.findByType('input');
 
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    inputElement.simulate('input', { target: { value: 'f' } });
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    const currentOptions = (componentRef.current as ComboBox).state.currentOptions;
-    expect(currentOptions.length).toEqual(DEFAULT_OPTIONS.length + 1);
-    expect(currentOptions[currentOptions.length - 1].text).toEqual('f');
+      simulateInputEvent(inputElement, 'f');
+      simulateKeydown(inputElement, KeyCodes.enter);
+
+      // open combobox
+      const buttonElement = container.root.findByType('button');
+      ReactTestUtils.act(() => {
+        buttonElement?.props?.onClick();
+      });
+
+      const options = container.root.findAll(
+        node =>
+          node.props.className &&
+          node.props.className.split(' ').includes('ms-ComboBox-option') &&
+          typeof node.type === 'string',
+      );
+
+      expect(options.length).toBe(DEFAULT_OPTIONS.length + 1);
+      expect(options[options.length - 1].props.title).toEqual('f');
+    });
   });
 
   it('Renders a default value with options', () => {
-    wrapper = mount(<ComboBox text="1" options={DEFAULT_OPTIONS} />);
-
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    expect(inputElement.props().value).toEqual('1');
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} text="1" />, container => {
+      const inputElement = container.root.findByType('input');
+      expect(inputElement.props.value).toEqual('1');
+    });
   });
 
   it('Renders a default value with no options', () => {
-    wrapper = mount(<ComboBox options={[]} text="1" />);
-
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    expect(inputElement.props().value).toEqual('1');
+    safeCreate(<ComboBox options={[]} text="1" />, container => {
+      const inputElement = container.root.findByType('input');
+      expect(inputElement.props.value).toEqual('1');
+    });
   });
 
   it('Can change items in uncontrolled case', () => {
-    domNode = renderIntoDocument(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS} />);
+    const ref = React.createRef<HTMLDivElement>();
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS} ref={ref} />, container => {
+      const buttonElement = ref.current?.querySelector('.ms-ComboBox button')!;
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.click(buttonElement);
+      });
 
-    const buttonElement = domNode.querySelector('.ms-ComboBox button')!;
-    ReactTestUtils.Simulate.click(buttonElement);
+      const secondItemElement = ref.current?.querySelector('.ms-ComboBox-option[data-index="1"]')!;
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.click(secondItemElement);
+      });
 
-    const secondItemElement = document.querySelector('.ms-ComboBox-option[data-index="1"]')!;
-    ReactTestUtils.Simulate.click(secondItemElement);
-
-    const inputElement = domNode.querySelector('.ms-ComboBox input') as HTMLInputElement;
-    expect(inputElement.value).toEqual('2');
+      const inputElement = ref.current?.querySelector('.ms-ComboBox input') as HTMLInputElement;
+      expect(inputElement.value).toEqual('2');
+    });
   });
 
   it('Does not automatically change items in controlled case', () => {
-    domNode = renderIntoDocument(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} />);
+    const ref = React.createRef<HTMLDivElement>();
+    safeCreate(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} ref={ref} />, container => {
+      const buttonElement = ref.current?.querySelector('.ms-ComboBox button')!;
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.click(buttonElement);
+      });
 
-    const buttonElement = domNode.querySelector('.ms-ComboBox button')!;
-    ReactTestUtils.Simulate.click(buttonElement);
+      const secondItemElement = ref.current?.querySelector('.ms-ComboBox-option[data-index="1"]')!;
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.click(secondItemElement);
+      });
 
-    const secondItemElement = document.querySelector('.ms-ComboBox-option[data-index="1"]')!;
-    ReactTestUtils.Simulate.click(secondItemElement);
-
-    const inputElement = domNode.querySelector('.ms-ComboBox input') as HTMLInputElement;
-    expect(inputElement.value).toEqual('1');
+      const inputElement = ref.current?.querySelector('.ms-ComboBox input') as HTMLInputElement;
+      expect(inputElement.value).toEqual('1');
+    });
   });
 
   it('Multiselect does not mutate props', () => {
-    domNode = renderIntoDocument(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} multiSelect />);
+    const ref = React.createRef<HTMLDivElement>();
+    safeCreate(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} multiSelect ref={ref} />, container => {
+      const buttonElement = ref.current?.querySelector('.ms-ComboBox button')!;
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.click(buttonElement);
+      });
 
-    const buttonElement = domNode.querySelector('.ms-ComboBox button')!;
-    ReactTestUtils.Simulate.click(buttonElement);
+      const buttons = ref.current?.querySelectorAll('.ms-ComboBox-option > input');
+      ReactTestUtils.act(() => {
+        ReactTestUtils.Simulate.change(buttons![1]);
+      });
 
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-    ReactTestUtils.Simulate.change(buttons[1]);
-
-    expect(!!DEFAULT_OPTIONS[1].selected).toEqual(false);
+      expect(!!DEFAULT_OPTIONS[1].selected).toEqual(false);
+    });
   });
 
   it('Can insert text in uncontrolled case with autoComplete and allowFreeform on', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={true} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'f');
+        expect(input.props.value).toBe('Foo');
+      },
     );
-
-    wrapper.find('input').simulate('input', { target: { value: 'f' } });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('Foo');
   });
 
   it('Can insert text in uncontrolled case with autoComplete on and allowFreeform off', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={false} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'f');
+        expect(input.props.value).toBe('Foo');
+      },
     );
-
-    wrapper.find('input').simulate('input', { target: { value: 'f' } });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('Foo');
   });
 
   it('Can insert non latin text in uncontrolled case with autoComplete on and allowFreeform off', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="0" options={RUSSIAN_OPTIONS} autoComplete="on" allowFreeform={false} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'п');
+        expect(input.props.value).toBe('папа');
+      },
     );
-
-    wrapper.find('input').simulate('input', { target: { value: 'п' } });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('папа');
   });
 
   it('Can insert text in uncontrolled case with autoComplete off and allowFreeform on', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={true} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'f');
+        expect(input.props.value).toBe('f');
+      },
     );
-    wrapper.find('input').simulate('input', { target: { value: 'f' } });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('f');
   });
 
   it('Can insert text in uncontrolled case with autoComplete and allowFreeform off', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={false} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'f');
+        expect(input.props.value).toBe('One');
+      },
     );
-    wrapper.find('input').simulate('keydown', { which: 'f' });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
   });
 
   it('Can insert an empty string in uncontrolled case with autoComplete and allowFreeform on', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={true} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, '');
+        simulateKeydown(input, KeyCodes.enter);
+        expect(input.props.value).toBe('');
+      },
     );
-    (wrapper.find('input').instance() as any).value = '';
-    wrapper.find('input').simulate('input', { target: { value: '' } });
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('');
   });
 
   it('Cannot insert an empty string in uncontrolled case with autoComplete on and allowFreeform off', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={false} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, '');
+        simulateKeydown(input, KeyCodes.enter);
+        expect(input.props.value).toBe('One');
+      },
     );
-
-    (wrapper.find('input').instance() as any).value = '';
-    wrapper.find('input').simulate('input', { target: { value: '' } });
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
   });
 
   it('Can insert an empty string in uncontrolled case with autoComplete off and allowFreeform on', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={true} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, '');
+        simulateKeydown(input, KeyCodes.enter);
+        expect(input.props.value).toBe('');
+      },
     );
-    (wrapper.find('input').instance() as any).value = '';
-    wrapper.find('input').simulate('input', { target: { value: '' } });
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('');
   });
 
   it('Cannot insert an empty string in uncontrolled case with autoComplete and allowFreeform off', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={false} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, '');
+        simulateKeydown(input, KeyCodes.enter);
+        expect(input.props.value).toBe('One');
+      },
     );
-    (wrapper.find('input').instance() as any).value = '';
-    wrapper.find('input').simulate('input', { target: { value: '' } });
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
   });
 
   // jeremy
@@ -345,17 +401,16 @@ describe('ComboBox', () => {
     'Can insert an empty string after removing a pending value in uncontrolled case ' +
       'with autoComplete and allowFreeform on',
     () => {
-      wrapper = mount(
+      safeCreate(
         <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={true} />,
+        container => {
+          const input = container.root.findByType('input');
+          simulateInputEvent(input, 'f');
+          simulateInputEvent(input, '');
+          simulateKeydown(input, KeyCodes.enter);
+          expect(input.props.value).toBe('');
+        },
       );
-
-      (wrapper.find('input').instance() as any).value = 'f';
-      wrapper.find('input').simulate('input', { target: { value: 'f' } });
-      (wrapper.find('input').instance() as any).value = '';
-      wrapper.find('input').simulate('input', { target: { value: '' } });
-      wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-      wrapper.update();
-      expect(wrapper.find('input').props().value).toEqual('');
     },
   );
 
@@ -363,17 +418,16 @@ describe('ComboBox', () => {
     'Cannot insert an empty string after removing a pending value in uncontrolled case ' +
       'with autoComplete on and allowFreeform off',
     () => {
-      wrapper = mount(
+      safeCreate(
         <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="on" allowFreeform={false} />,
+        container => {
+          const input = container.root.findByType('input');
+          simulateInputEvent(input, 'f');
+          simulateInputEvent(input, '');
+          simulateKeydown(input, KeyCodes.enter);
+          expect(input.props.value).toBe('Foo');
+        },
       );
-
-      (wrapper.find('input').instance() as any).value = 'f';
-      wrapper.find('input').simulate('input', { target: { value: 'f' } });
-      (wrapper.find('input').instance() as any).value = '';
-      wrapper.find('input').simulate('input', { target: { value: '' } });
-      wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-      wrapper.update();
-      expect(wrapper.find('input').props().value).toEqual('Foo');
     },
   );
 
@@ -381,17 +435,16 @@ describe('ComboBox', () => {
     'Can insert an empty string after removing a pending value in uncontrolled case ' +
       'with autoComplete off and allowFreeform on',
     () => {
-      wrapper = mount(
+      safeCreate(
         <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={true} />,
+        container => {
+          const input = container.root.findByType('input');
+          simulateInputEvent(input, 'f');
+          simulateInputEvent(input, '');
+          simulateKeydown(input, KeyCodes.enter);
+          expect(input.props.value).toBe('');
+        },
       );
-
-      (wrapper.find('input').instance() as any).value = 'f';
-      wrapper.find('input').simulate('input', { target: { value: 'f' } });
-      (wrapper.find('input').instance() as any).value = '';
-      wrapper.find('input').simulate('input', { target: { value: '' } });
-      wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-      wrapper.update();
-      expect(wrapper.find('input').props().value).toEqual('');
     },
   );
 
@@ -399,111 +452,127 @@ describe('ComboBox', () => {
     'Cannot insert an empty string after removing a pending value in uncontrolled case ' +
       'with autoComplete and allowFreeform off',
     () => {
-      wrapper = mount(
+      safeCreate(
         <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} autoComplete="off" allowFreeform={false} />,
+        container => {
+          const input = container.root.findByType('input');
+          simulateInputEvent(input, 'f');
+          simulateInputEvent(input, '');
+          simulateKeydown(input, KeyCodes.enter);
+          expect(input.props.value).toBe('One');
+        },
       );
-      (wrapper.find('input').instance() as any).value = 'f';
-      wrapper.find('input').simulate('input', { target: { value: 'f' } });
-      (wrapper.find('input').instance() as any).value = '';
-      wrapper.find('input').simulate('input', { target: { value: '' } });
-      wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-      wrapper.update();
-      expect(wrapper.find('input').props().value).toEqual('One');
     },
   );
 
   it('Can change selected option with keyboard', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.down });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('Foo');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.down);
+      expect(input.props.value).toEqual('Foo');
+    });
   });
 
   it('Can change selected option with keyboard, looping from top to bottom', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.up });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('Bar');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.up);
+      expect(input.props.value).toEqual('Bar');
+    });
   });
 
   it('Can change selected option with keyboard, looping from bottom to top', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="3" options={DEFAULT_OPTIONS2} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.down });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
+    safeCreate(<ComboBox defaultSelectedKey="3" options={DEFAULT_OPTIONS2} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.down);
+      expect(input.props.value).toEqual('One');
+    });
   });
 
   it('Can change selected option with keyboard, looping from top to bottom', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS3} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.up });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('Bar');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS3} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.up);
+      expect(input.props.value).toEqual('Bar');
+    });
   });
 
   it('Cannot insert text while disabled', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.a });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.a);
+      expect(input.props.value).toEqual('One');
+    });
   });
 
   it('Cannot change selected option with keyboard while disabled', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.down });
-    wrapper.update();
-    expect(wrapper.find('input').props().value).toEqual('One');
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />, container => {
+      const input = container.root.findByType('input');
+      simulateKeydown(input, KeyCodes.down);
+      expect(input.props.value).toEqual('One');
+    });
   });
 
   it('Cannot expand the menu when clicking on the input while disabled', () => {
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS2} disabled={true} />);
-    wrapper.find('input').simulate('click');
-    expect(wrapper.find('.is-opened').length).toEqual(0);
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS2} disabled={true} />, container => {
+      const input = container.root.findByType('input');
+      input.props.onClick({});
+      expect(findNodeWithClass(container, 'is-opened', true).length).toEqual(0);
+    });
   });
 
   it('Cannot expand the menu when clicking on the button while disabled', () => {
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />);
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const buttonElement = wrapper.find('button');
-    buttonElement.simulate('click');
-    expect(comboBoxRoot.find('.is-opened').length).toEqual(0);
+    safeCreate(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} disabled={true} />, container => {
+      const buttonElement = container.root.findByType('button');
+      buttonElement.props.onClick({});
+      expect(findNodeWithClass(container, 'is-opened', true).length).toEqual(0);
+    });
   });
 
   it('Call onMenuOpened when clicking on the button', () => {
     const onMenuOpenMock = jest.fn();
 
-    wrapper = mount(<ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} onMenuOpen={onMenuOpenMock} />);
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const buttonElement = comboBoxRoot.find('button');
-    buttonElement.simulate('click');
-    expect(onMenuOpenMock.mock.calls.length).toBe(1);
+    safeCreate(
+      <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} onMenuOpen={onMenuOpenMock} />,
+      container => {
+        const buttonElement = container.root.findByType('button');
+        buttonElement.props.onClick({});
+        expect(onMenuOpenMock.mock.calls.length).toBe(1);
+      },
+    );
   });
 
   it('Opens on focus when openOnKeyboardFocus is true', () => {
     const onMenuOpenMock = jest.fn();
 
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" openOnKeyboardFocus options={DEFAULT_OPTIONS2} onMenuOpen={onMenuOpenMock} />,
+      container => {
+        const input = container.root.findByType('input');
+
+        input.props.onFocus?.();
+        input.props.onKeyUp?.({});
+        expect(onMenuOpenMock.mock.calls.length).toBe(1);
+      },
     );
-    const comboBoxRoot = wrapper.find('.ms-ComboBox-Input').find('input');
-    comboBoxRoot.simulate('focus');
-    comboBoxRoot.simulate('keyup');
-    expect(onMenuOpenMock.mock.calls.length).toBe(1);
   });
 
   it('Call onMenuOpened when touch start on the input', () => {
-    wrapper = mount(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS2} onMenuOpen={returnUndefined} allowFreeform={true} />,
+      container => {
+        const input = container.root.findByType('input');
+
+        input.props.onTouchStart?.();
+        input.props.onClick?.();
+
+        expect(findNodeWithClass(container, 'is-open', true).length).toEqual(1);
+      },
+      {
+        createNodeMock: element =>
+          element.type === 'div' ? { addEventListener: jest.fn(), removeEventListener: jest.fn() } : undefined,
+      },
     );
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input');
-
-    // in a normal scenario, when we do a touchstart we would also cause a
-    // click event to fire. This doesn't happen in the simulator so we're
-    // manually adding this in.
-    inputElement.simulate('touchstart');
-    inputElement.simulate('click');
-
-    expect(wrapper.find('.is-open').length).toEqual(1);
   });
 
   it('onPendingValueChanged triggers for all indexes', () => {
@@ -513,20 +582,23 @@ describe('ComboBox', () => {
         indexSeen.push(index);
       }
     };
-    wrapper = mount(
+    safeCreate(
       <ComboBox
         options={DEFAULT_OPTIONS}
         defaultSelectedKey="1"
         allowFreeform={true}
         onPendingValueChanged={pendingValueChangedHandler}
       />,
+      container => {
+        const input = container.root.findByType('input');
+
+        simulateInputEvent(input, 'f');
+        simulateKeydown(input, KeyCodes.down);
+        simulateKeydown(input, KeyCodes.up);
+        expect(indexSeen).toContain(0);
+        expect(indexSeen).toContain(1);
+      },
     );
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    inputElement.simulate('input', { target: { value: 'f' } });
-    inputElement.simulate('keydown', { which: KeyCodes.down });
-    inputElement.simulate('keydown', { which: KeyCodes.up });
-    expect(indexSeen).toContain(0);
-    expect(indexSeen).toContain(1);
   });
 
   it('onPendingValueChanged is called with an empty string when the input is cleared', () => {
@@ -535,89 +607,83 @@ describe('ComboBox', () => {
       changedValue = value;
     };
 
-    const baseNode = document.createElement('div');
-    document.body.appendChild(baseNode);
-
-    const component = ReactDOM.render(
+    safeCreate(
       <ComboBox options={DEFAULT_OPTIONS} allowFreeform={true} onPendingValueChanged={pendingValueChangedHandler} />,
-      baseNode,
+      container => {
+        const input = container.root.findByType('input');
+
+        simulateInputEvent(input, 'a');
+        expect(changedValue).toEqual('a');
+
+        simulateInputEvent(input, '');
+        expect(changedValue).toEqual('');
+      },
     );
-
-    const input = (ReactDOM.findDOMNode((component as unknown) as React.ReactInstance) as Element).querySelector(
-      'input',
-    ) as HTMLInputElement;
-    if (input === null) {
-      throw new Error('ComboBox input element is null');
-    }
-
-    // Simulate typing one character into the ComboBox input
-    input.value = 'a';
-    ReactTestUtils.Simulate.input(input);
-    expect(changedValue).toEqual('a');
-
-    // Simulate clearing the ComboBox input
-    input.value = '';
-    ReactTestUtils.Simulate.input(input);
-    expect(changedValue).toEqual('');
   });
 
   it('suggestedDisplayValue is called undefined when the selected input is cleared', () => {
-    const componentRef = React.createRef<any>();
-    wrapper = mount(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} componentRef={componentRef} />);
+    safeCreate(<ComboBox selectedKey="1" options={DEFAULT_OPTIONS} />, container => {
+      const input = container.root.findByType('input');
+      expect(input.props.value).toEqual('1');
 
-    // SelectedKey is still the same
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    expect(inputElement.props().value).toEqual('1');
-
-    // SelectedKey is set to null
-    wrapper.setProps({ selectedKey: null });
-    expect(wrapper.find('input').props().value).toEqual('');
-
-    const suggestedDisplay = (componentRef.current as ComboBox).state.suggestedDisplayValue;
-    expect(suggestedDisplay).toEqual(undefined);
+      renderer.act(() => {
+        container.update(<ComboBox selectedKey={null} options={DEFAULT_OPTIONS} />);
+      });
+      expect(input.props.value).toEqual('');
+    });
   });
 
   it('Can type a complete option with autocomplete and allowFreeform on and submit it', () => {
-    let updatedOption;
-    let updatedIndex;
+    let updatedOption: IComboBoxOption | undefined;
+    let updatedIndex: number | undefined;
     const onChange = jest.fn((event: React.FormEvent<IComboBox>, option?: IComboBoxOption, index?: number) => {
       updatedOption = option;
       updatedIndex = index;
     });
     const initialOption = { key: '1', text: 'Text' };
 
-    wrapper = mount(<ComboBox options={[initialOption]} autoComplete="on" allowFreeform={true} onChange={onChange} />);
-    const inputElement: InputElementWrapper = wrapper.find('input');
-    inputElement.simulate('input', { target: { value: 't' } });
-    inputElement.simulate('input', { target: { value: 'e' } });
-    inputElement.simulate('input', { target: { value: 'x' } });
-    inputElement.simulate('input', { target: { value: 't' } });
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    expect(onChange).toHaveBeenCalledTimes(1);
-    expect(updatedOption).toEqual(initialOption);
-    expect(updatedIndex).toEqual(0);
+    safeCreate(
+      <ComboBox options={[initialOption]} autoComplete="on" allowFreeform={true} onChange={onChange} />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 't');
+        simulateInputEvent(input, 'te');
+        simulateInputEvent(input, 'tex');
+        simulateInputEvent(input, 'text');
+        simulateKeydown(input, KeyCodes.enter);
 
-    wrapper.update();
-    expect(wrapper.find('.ms-ComboBox input').props().value).toEqual('Text');
+        expect(onChange).toHaveBeenCalledTimes(1);
+        expect(updatedOption).toEqual(initialOption);
+        expect(updatedIndex).toEqual(0);
+
+        expect(input.props.value).toEqual('Text');
+      },
+    );
   });
 
   it('merges callout classNames', () => {
-    domNode = renderIntoDocument(
+    safeCreate(
       <ComboBox defaultSelectedKey="1" options={DEFAULT_OPTIONS} calloutProps={{ className: 'foo' }} />,
+      container => {
+        const buttonElement = container.root.find(
+          node => node.type === 'button' && node.props?.className?.includes?.('ms-ComboBox'),
+        );
+
+        ReactTestUtils.act(() => {
+          buttonElement?.props?.onClick?.();
+        });
+
+        const callout = container.root.find(node => node.props?.className?.split?.(' ').includes?.('ms-Callout'));
+        expect(callout).toBeDefined();
+        expect(callout.props.className.includes('ms-ComboBox-callout')).toBeTruthy();
+        expect(callout.props.className.includes('foo')).toBeTruthy();
+      },
     );
-
-    const buttonElement = domNode.querySelector('.ms-ComboBox button')!;
-    ReactTestUtils.Simulate.click(buttonElement);
-
-    const callout = document.querySelector('.ms-Callout')!;
-    expect(callout).toBeDefined();
-    expect(callout.classList.contains('ms-ComboBox-callout')).toBeTruthy();
-    expect(callout.classList.contains('foo')).toBeTruthy();
   });
 
   it('Can clear text in controlled case with autoComplete off and allowFreeform on', () => {
-    let updatedText;
-    wrapper = mount(
+    let updatedText: string | undefined;
+    safeCreate(
       <ComboBox
         options={DEFAULT_OPTIONS}
         autoComplete="off"
@@ -627,20 +693,19 @@ describe('ComboBox', () => {
           updatedText = value;
         }}
       />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, '');
+        simulateKeydown(input, KeyCodes.enter);
+
+        expect(updatedText).toEqual('');
+      },
     );
-
-    const input = wrapper.find('input');
-    (input.instance() as any).value = '';
-    input.simulate('input', { target: { value: '' } });
-    input.simulate('keydown', { which: KeyCodes.enter });
-    wrapper.update();
-
-    expect(updatedText).toEqual('');
   });
 
   it('Can clear text in controlled case with autoComplete off and allowFreeform on', () => {
-    let updatedText;
-    wrapper = mount(
+    let updatedText: string | undefined;
+    safeCreate(
       <ComboBox
         options={DEFAULT_OPTIONS}
         autoComplete="off"
@@ -649,173 +714,183 @@ describe('ComboBox', () => {
           updatedText = value;
         }}
       />,
+      container => {
+        const input = container.root.findByType('input');
+        simulateInputEvent(input, 'ab');
+        simulateKeydown(input, KeyCodes.backspace);
+        simulateInputEvent(input, 'a');
+        simulateKeydown(input, KeyCodes.backspace);
+        simulateInputEvent(input, '');
+        expect(input.props.value).toEqual('');
+        simulateKeydown(input, KeyCodes.enter);
+        expect(updatedText).toEqual('');
+      },
     );
-
-    const input = wrapper.find('input');
-    (input.instance() as any).value = 'ab';
-    input.simulate('input', { target: { value: 'ab' } });
-    input.simulate('keydown', { which: KeyCodes.backspace });
-    input.simulate('input', { target: { value: 'a' } });
-    input.simulate('keydown', { which: KeyCodes.backspace });
-    wrapper.update();
-
-    (input.instance() as any).value = '';
-    input.simulate('input', { target: { value: '' } });
-    wrapper.update();
-    expect((input.instance() as any).value).toEqual('');
-    input.simulate('keydown', { which: KeyCodes.enter });
-
-    expect(updatedText).toEqual('');
   });
 
-  it('in multiSelect mode, selectedIndices are correct after performing multiple selections using mouse click', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(<ComboBox multiSelect options={DEFAULT_OPTIONS} componentRef={comboBoxRef} />);
+  //it('in multiSelect mode, selectedIndices are correct after performing multiple selections using mouse click', () =>
+  // {
+  //   const comboBoxRef = React.createRef<IComboBox>();
+  //   wrapper = mount(<ComboBox multiSelect options={DEFAULT_OPTIONS} componentRef={comboBoxRef} />);
 
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input');
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
+  //   const comboBoxRoot = wrapper.find('.ms-ComboBox');
+  //   const inputElement = comboBoxRoot.find('input');
+  //   inputElement.simulate('keydown', { which: KeyCodes.enter });
+  //   const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
 
-    ReactTestUtils.Simulate.change(buttons[0]);
-    ReactTestUtils.Simulate.change(buttons[2]);
-    ReactTestUtils.Simulate.change(buttons[1]);
+  //   ReactTestUtils.Simulate.change(buttons[0]);
+  //   ReactTestUtils.Simulate.change(buttons[2]);
+  //   ReactTestUtils.Simulate.change(buttons[1]);
 
-    expect((comboBoxRef.current as ComboBox).state.selectedIndices).toEqual([0, 2, 1]);
-  });
+  //   expect(((comboBoxRef.current as unknown) as ComboBox).state.selectedIndices).toEqual([0, 2, 1]);
+  // });
 
   it('in multiSelect mode, defaultselected keys produce correct display input', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(
+    safeCreate(
       <ComboBox
         multiSelect
         options={DEFAULT_OPTIONS}
-        componentRef={comboBoxRef}
         selectedKey={[DEFAULT_OPTIONS[0].key as string, DEFAULT_OPTIONS[2].key as string]}
       />,
+      container => {
+        const comboBoxRoot = findNodeWithClass(container, 'ms-ComboBox');
+        const inputElement = comboBoxRoot.findByType('input');
+        const caretElement = findNodeWithClass(container, 'ms-ComboBox-CaretDown-button');
+        renderer.act(() => {
+          caretElement.props?.onClick?.();
+        });
+        renderer.act(() => {
+          inputElement.props.onBlur?.({});
+        });
+        const button = findNodeWithClass(container, 'ms-ComboBox-option', true)
+          .filter(node => node.props.className.includes('ms-Checkbox'))
+          .map(node => node.findByType('input'));
+        renderer.act(() => {
+          button[2]?.props?.onChange?.({ persist: jest.fn() });
+        });
+        renderer.act(() => {
+          button[0]?.props?.onChange?.({ persist: jest.fn() });
+        });
+        const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].map(({ text }) => text).join(', ');
+
+        expect(inputElement.props.value).toEqual(compare);
+      },
     );
-
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input');
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-
-    ReactTestUtils.Simulate.change(buttons[2]);
-    ReactTestUtils.Simulate.change(buttons[0]);
-    const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].reduce((previous: string, current: IComboBoxOption) => {
-      if (previous !== '') {
-        return previous + ', ' + current.text;
-      }
-      return current.text;
-    }, '');
-
-    expect((inputElement.instance() as any).value).toEqual(compare);
   });
 
   it('in multiSelect mode, input has correct value', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(<ComboBox multiSelect options={DEFAULT_OPTIONS} componentRef={comboBoxRef} />);
+    safeCreate(<ComboBox multiSelect options={DEFAULT_OPTIONS} />, container => {
+      const comboBoxRoot = findNodeWithClass(container, 'ms-ComboBox');
+      const inputElement = comboBoxRoot.findByType('input');
+      const caretElement = findNodeWithClass(container, 'ms-ComboBox-CaretDown-button');
+      renderer.act(() => {
+        caretElement.props?.onClick?.();
+      });
+      renderer.act(() => {
+        inputElement.props.onBlur?.({});
+      });
+      const button = findNodeWithClass(container, 'ms-ComboBox-option', true)
+        .filter(node => node.props.className.includes('ms-Checkbox'))
+        .map(node => node.findByType('input'));
+      renderer.act(() => {
+        button[0]?.props?.onChange?.({ persist: jest.fn() });
+      });
+      renderer.act(() => {
+        button[2]?.props?.onChange?.({ persist: jest.fn() });
+      });
+      const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].map(({ text }) => text).join(', ');
 
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input');
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-
-    ReactTestUtils.Simulate.change(buttons[0]);
-    ReactTestUtils.Simulate.change(buttons[2]);
-    const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].reduce((previous: string, current: IComboBoxOption) => {
-      if (previous !== '') {
-        return previous + ', ' + current.text;
-      }
-      return current.text;
-    }, '');
-
-    expect((inputElement.instance() as any).value).toEqual(compare);
+      expect(inputElement.props.value).toEqual(compare);
+    });
   });
 
   it('in multiSelect mode, input has correct value when multiSelectDelimiter specified', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(
-      <ComboBox multiSelect multiSelectDelimiter="; " options={DEFAULT_OPTIONS} componentRef={comboBoxRef} />,
-    );
+    safeCreate(<ComboBox multiSelect multiSelectDelimiter="; " options={DEFAULT_OPTIONS} />, container => {
+      const comboBoxRoot = findNodeWithClass(container, 'ms-ComboBox');
+      const inputElement = comboBoxRoot.findByType('input');
+      const caretElement = findNodeWithClass(container, 'ms-ComboBox-CaretDown-button');
+      renderer.act(() => {
+        caretElement.props?.onClick?.();
+      });
+      renderer.act(() => {
+        inputElement.props.onBlur?.({});
+      });
+      const button = findNodeWithClass(container, 'ms-ComboBox-option', true)
+        .filter(node => node.props.className.includes('ms-Checkbox'))
+        .map(node => node.findByType('input'));
+      renderer.act(() => {
+        button[0]?.props?.onChange?.({ persist: jest.fn() });
+      });
+      renderer.act(() => {
+        button[2]?.props?.onChange?.({ persist: jest.fn() });
+      });
+      const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].map(({ text }) => text).join('; ');
 
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input');
-    inputElement.simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-
-    ReactTestUtils.Simulate.change(buttons[0]);
-    ReactTestUtils.Simulate.change(buttons[2]);
-    const compare = [DEFAULT_OPTIONS[0], DEFAULT_OPTIONS[2]].reduce((previous: string, current: IComboBoxOption) => {
-      if (previous !== '') {
-        return previous + '; ' + current.text;
-      }
-      return current.text;
-    }, '');
-
-    expect((inputElement.instance() as any).value).toEqual(compare);
+      expect(inputElement.props.value).toEqual(compare);
+    });
   });
 
   it('in multiSelect mode, optional onItemClick callback invoked per option select', () => {
     const onItemClickMock = jest.fn();
-    wrapper = mount(<ComboBox multiSelect options={DEFAULT_OPTIONS} onItemClick={onItemClickMock} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-
-    ReactTestUtils.Simulate.change(buttons[0]);
-    ReactTestUtils.Simulate.change(buttons[1]);
-    ReactTestUtils.Simulate.change(buttons[2]);
-
-    expect(onItemClickMock).toHaveBeenCalledTimes(3);
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} onItemClick={onItemClickMock} />, container => {
+      const caretElement = findNodeWithClass(container, 'ms-ComboBox-CaretDown-button');
+      renderer.act(() => {
+        caretElement?.props?.onClick?.();
+      });
+      const button = findNodeWithClass(container, 'ms-ComboBox-option', true);
+      renderer.act(() => {
+        button[0]?.props?.onClick?.({ persist: jest.fn() });
+        button[1]?.props?.onClick?.({ persist: jest.fn() });
+        button[2]?.props?.onClick?.({ persist: jest.fn() });
+      });
+      expect(onItemClickMock).toHaveBeenCalledTimes(3);
+    });
   });
 
   it('invokes optional onItemClick callback on option select', () => {
     const onItemClickMock = jest.fn();
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} onItemClick={onItemClickMock} />);
-    wrapper.find('input').simulate('keydown', { which: KeyCodes.enter });
-    const buttons = document.querySelectorAll('.ms-ComboBox-option');
-
-    (buttons[0] as HTMLInputElement).click();
-
-    expect(onItemClickMock).toHaveBeenCalledTimes(1);
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} onItemClick={onItemClickMock} />, container => {
+      const caretElement = findNodeWithClass(container, 'ms-ComboBox-CaretDown-button');
+      ReactTestUtils.act(() => {
+        caretElement?.props?.onClick?.();
+      });
+      const button = findNodeWithClass(container, 'ms-ComboBox-option', true);
+      renderer.act(() => {
+        button?.[0]?.props?.onClick?.({ persist: jest.fn() });
+      });
+      expect(onItemClickMock).toHaveBeenCalledTimes(1);
+    });
   });
 
   it('allows adding a custom aria-describedby id to the input', () => {
-    const comboBoxRef = React.createRef<any>();
     const customId = 'customAriaDescriptionId';
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} componentRef={comboBoxRef} ariaDescribedBy={customId} />);
-
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input').getDOMNode();
-    const ariaDescribedByAttribute = inputElement.getAttribute('aria-describedby');
-    expect(ariaDescribedByAttribute).toMatch(new RegExp('\\b' + customId + '\\b'));
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} ariaDescribedBy={customId} />, container => {
+      const inputElement = container.root.findByType('input');
+      expect((inputElement.props as React.HTMLAttributes<HTMLInputElement>)['aria-describedby']).toMatch(
+        new RegExp('\\b' + customId + '\\b'),
+      );
+    });
   });
 
   it('adds aria-required to the DOM when the required prop is set to true', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} componentRef={comboBoxRef} required={true} />);
-
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input').getDOMNode();
-    const ariaRequiredAttribute = inputElement.getAttribute('aria-required');
-    expect(ariaRequiredAttribute).toEqual('true');
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} required={true} />, container => {
+      const inputElement = container.root.findByType('input');
+      expect((inputElement.props as React.HTMLAttributes<HTMLInputElement>)['aria-required']).toBe(true);
+    });
   });
 
   it('does not add aria-required to the DOM when the required prop is not set', () => {
-    const comboBoxRef = React.createRef<any>();
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} componentRef={comboBoxRef} />);
-
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
-    const inputElement = comboBoxRoot.find('input').getDOMNode();
-    const ariaRequiredAttribute = inputElement.getAttribute('aria-required');
-    expect(ariaRequiredAttribute).toBeNull();
+    safeCreate(<ComboBox options={DEFAULT_OPTIONS} />, container => {
+      const inputElement = container.root.findByType('input');
+      expect((inputElement.props as React.HTMLAttributes<HTMLInputElement>)['aria-required']).toBeUndefined();
+    });
   });
 
   it('test persistMenu, callout should exist before and after opening menu', () => {
     const onMenuOpenMock = jest.fn();
     const onMenuDismissedMock = jest.fn();
 
-    wrapper = mount(
+    safeCreate(
       <ComboBox
         defaultSelectedKey="1"
         persistMenu={true}
@@ -823,144 +898,64 @@ describe('ComboBox', () => {
         onMenuOpen={onMenuOpenMock}
         onMenuDismissed={onMenuDismissedMock}
       />,
-    );
-    const comboBoxRoot = wrapper.find('.ms-ComboBox');
+      container => {
+        const comboBoxRoot = findNodeWithClass(container, 'ms-ComboBox');
 
-    // Find menu
-    const calloutBeforeOpen = document.querySelector('.ms-Callout')!;
-    expect(calloutBeforeOpen).toBeDefined();
-    expect(calloutBeforeOpen.classList.contains('ms-ComboBox-callout')).toBeTruthy();
+        // Find menu
+        const calloutBeforeOpen = findNodeWithClass(container, 'ms-Callout');
+        expect(calloutBeforeOpen).toBeDefined();
+        expect(calloutBeforeOpen?.props?.className?.includes?.('ms-ComboBox-callout')).toBeTruthy();
 
-    // Open combobox
-    const buttonElement = comboBoxRoot.find('button');
-    buttonElement.simulate('click');
-    expect(onMenuOpenMock.mock.calls.length).toBe(1);
+        // Open combobox
+        const buttonElement = comboBoxRoot.findByType('button');
+        ReactTestUtils.act(() => {
+          buttonElement?.props?.onClick();
+        });
+        expect(onMenuOpenMock.mock.calls.length).toBe(1);
 
-    // Close combobox
-    buttonElement.simulate('click');
-    expect(onMenuDismissedMock.mock.calls.length).toBe(1);
+        // Close combobox
+        ReactTestUtils.act(() => {
+          buttonElement?.props?.onClick();
+        });
+        expect(onMenuDismissedMock.mock.calls.length).toBe(1);
 
-    // Ensure menu is still there
-    const calloutAfterClose = document.querySelector('.ms-Callout')!;
-    expect(calloutAfterClose).toBeDefined();
-    expect(calloutAfterClose.classList.contains('ms-ComboBox-callout')).toBeTruthy();
-  });
-
-  // Adds currentPendingValue to options and makes it selected onBlur
-  // if allowFreeFrom is true for multiselect with default selected values
-  it('adds currentPendingValue to options and selects if multiSelected with default values', () => {
-    const componentRef = React.createRef<any>();
-    const comboBoxOption: IComboBoxOption = {
-      key: 'ManuallyEnteredValue',
-      text: 'ManuallyEnteredValue',
-      selected: true,
-    };
-    wrapper = mount(
-      <ComboBox
-        multiSelect
-        options={DEFAULT_OPTIONS}
-        defaultSelectedKey={['1', '2', '3']}
-        allowFreeform={true}
-        componentRef={componentRef}
-      />,
-    );
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS], [0, 1, 2]);
-    inputElement.simulate('focus');
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS], [0, 1, 2]);
-    inputElement.simulate('input', { target: { value: 'ManuallyEnteredValue' } });
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS], [0, 1, 2]);
-    inputElement.simulate('blur');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [0, 1, 2, 3]);
-  });
-
-  // Adds currentPendingValue to options and makes it selected onBlur
-  // if allowFreeForm is true for multiSelect with no default value selected
-  it('adds currentPendingValue to options and selects for multiSelect with no default value', () => {
-    const componentRef = React.createRef<any>();
-    const comboBoxOption: IComboBoxOption = {
-      key: 'ManuallyEnteredValue',
-      text: 'ManuallyEnteredValue',
-      selected: true,
-    };
-    wrapper = mount(
-      <ComboBox multiSelect options={DEFAULT_OPTIONS} allowFreeform={true} componentRef={componentRef} />,
-    );
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('focus');
-    inputElement.simulate('keyup', { which: 10 });
-    expect((componentRef.current as ComboBox).state.focusState).toEqual('focused');
-    _verifyStateVariables(componentRef, 'focused', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('input', { target: { value: 'ManuallyEnteredValue' } });
-    _verifyStateVariables(componentRef, 'focused', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('blur');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [3]);
-
-    inputElement.simulate('focus');
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [3]);
-    inputElement.simulate('input', { target: { value: 'ManuallyEnteredValue' } });
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [3]);
-
-    // This should toggle the checkbox off. With multi-select the currentPendingValue is not reset on input change
-    // because it would break keyboard accessibility
-    wrapper.find('.ms-ComboBox button').simulate('click');
-    const buttons = document.querySelectorAll('.ms-ComboBox-option > input');
-    ReactTestUtils.Simulate.change(buttons[3]);
-
-    // with 'ManuallyEnteredValue' still in the input, on blur it should toggle the check back to on
-    inputElement.simulate('blur');
-    _verifyStateVariables(
-      componentRef,
-      'none',
-      [
-        ...DEFAULT_OPTIONS,
-        {
-          ...comboBoxOption,
-          selected: true,
-        },
-      ],
-      [3],
+        // Ensure menu is still there
+        const calloutAfterClose = findNodeWithClass(container, 'ms-Callout');
+        expect(calloutAfterClose).toBeDefined();
+        expect(calloutBeforeOpen?.props?.className?.includes?.('ms-ComboBox-callout')).toBeTruthy();
+      },
     );
   });
-
-  // adds currentPendingValue to options and makes it selected onBlur
-  // if allowFreeForm is true for singleSelect
-  it('adds currentPendingValue to options and selects for singleSelect', () => {
-    const componentRef = React.createRef<any>();
-    const comboBoxOption: IComboBoxOption = {
-      key: 'ManuallyEnteredValue',
-      text: 'ManuallyEnteredValue',
-    };
-    wrapper = mount(<ComboBox options={DEFAULT_OPTIONS} allowFreeform={true} componentRef={componentRef} />);
-    const inputElement: InputElementWrapper = wrapper.find('.ms-ComboBox input');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('focus');
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('input', { target: { value: 'ManuallyEnteredValue' } });
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS], []);
-    inputElement.simulate('blur');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [3]);
-
-    inputElement.simulate('focus');
-    _verifyStateVariables(componentRef, 'focusing', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [3]);
-    const buttonElement: any = wrapper.find('.ms-ComboBox button')! as any;
-    buttonElement.simulate('click');
-    const secondItem = document.querySelector('.ms-ComboBox-option[data-index="2"]')!;
-    ReactTestUtils.Simulate.click(secondItem);
-
-    inputElement.simulate('blur');
-    _verifyStateVariables(componentRef, 'none', [...DEFAULT_OPTIONS, { ...comboBoxOption }], [2]);
-  });
-
-  function _verifyStateVariables(
-    componentRef: React.RefObject<any>,
-    focusState: 'none' | 'focused' | 'focusing',
-    currentOptions: IComboBoxOption[],
-    selectedIndices?: number[],
-  ): void {
-    expect((componentRef.current as ComboBox).state.focusState).toEqual(focusState);
-    expect((componentRef.current as ComboBox).state.selectedIndices).toEqual(selectedIndices);
-    expect((componentRef.current as ComboBox).state.currentOptions).toEqual(currentOptions);
-  }
 });
+
+function simulateInputEvent(input: renderer.ReactTestInstance, value: string) {
+  renderer.act(() => {
+    input.props.onInput?.({ target: { value }, nativeEvent: { isComposing: false } });
+  });
+}
+
+function simulateKeydown(input: renderer.ReactTestInstance, which: KeyCodes) {
+  renderer.act(() => {
+    input.props.onKeyDown?.({
+      which,
+      nativeEvent: { isComposing: false },
+      stopPropagation: jest.fn(),
+      preventDefault: jest.fn(),
+      persist: jest.fn(),
+    });
+  });
+}
+
+function findNodeWithClass(container: renderer.ReactTestRenderer, className: string): renderer.ReactTestInstance;
+function findNodeWithClass(
+  container: renderer.ReactTestRenderer,
+  className: string,
+  findAll: true,
+): renderer.ReactTestInstance[];
+function findNodeWithClass(
+  container: renderer.ReactTestRenderer,
+  className: string,
+  findAll?: true,
+): renderer.ReactTestInstance | renderer.ReactTestInstance[] {
+  return container.root[findAll ? 'findAll' : 'find'](node => node.props?.className?.split(' ')?.includes?.(className));
+}
