+++ packages/react/src/components/ComboBox/ComboBox.tsx
@@ -18,41 +18,26 @@ import {
   warnMutuallyExclusive,
   Async,
   EventGroup,
+  getPropsWithDefaults,
 } from '../../Utilities';
-import { Callout } from '../../Callout';
+import { Callout, DirectionalHint } from '../../Callout';
 import { Checkbox } from '../../Checkbox';
-import { CommandButton, IButtonStyles, IconButton } from '../../Button';
-import { DirectionalHint } from '@fluentui/react-internal/lib/common/DirectionalHint';
 import { getCaretDownButtonStyles, getOptionStyles, getStyles } from './ComboBox.styles';
-import { getClassNames, getComboBoxOptionClassNames } from './ComboBox.classNames';
-import {
-  IComboBoxOption,
-  IComboBoxOptionStyles,
-  IComboBoxProps,
-  IOnRenderComboBoxLabelProps,
-  IComboBoxClassNames,
-} from './ComboBox.types';
+import { getClassNames, getComboBoxOptionClassNames, IComboBoxClassNames } from './ComboBox.classNames';
+import { IComboBoxOption, IComboBoxOptionStyles, IComboBoxProps, IOnRenderComboBoxLabelProps } from './ComboBox.types';
 import { KeytipData } from '../../KeytipData';
 import { Label } from '../../Label';
 import { SelectableOptionMenuItemType, getAllSelectedOptions } from '../../SelectableOption';
-import { BaseButton, Button } from '@fluentui/react-internal/lib/Button';
+import { BaseButton, Button, CommandButton, IButtonStyles, IconButton } from '../../compat/Button';
 import { ICalloutProps } from '../../Callout';
+import { useMergedRefs } from '@uifabric/react-hooks';
 
 export interface IComboBoxState {
   /** The open state */
   isOpen?: boolean;
 
-  /** The currently selected indices */
-  selectedIndices?: number[];
-
-  /** The focused state of the comboBox */
+  /** The focused state of the combo box */
   focusState?: 'none' | 'focused' | 'focusing';
-
-  /** This value is used for the autocomplete hint value */
-  suggestedDisplayValue?: string;
-
-  /** The options currently available for the callout */
-  currentOptions: IComboBoxOption[];
 
   /**
    * When taking input, this will store the index that the options input matches
@@ -108,35 +93,115 @@ interface IComboBoxOptionWrapperProps extends IComboBoxOption {
 }
 
 /**
- * Internal class that is used to wrap all ComboBox options.
- * This is used to customize when we want to rerender components,
- * so we don't rerender every option every time render is executed.
+ * Internal component that is used to wrap all ComboBox options.
+ * This is used to customize when we want to re-render components,
+ * so we don't re-render every option every time render is executed.
  */
-class ComboBoxOptionWrapper extends React.Component<IComboBoxOptionWrapperProps, {}> {
-  public render(): React.ReactNode {
-    return this.props.render();
-  }
-
-  public shouldComponentUpdate(newProps: IComboBoxOptionWrapperProps): boolean {
+const ComboBoxOptionWrapper = React.memo(
+  ({ render }: IComboBoxOptionWrapperProps) => render(),
+  (
+    { render: oldRender, ...oldProps }: IComboBoxOptionWrapperProps,
+    { render: newRender, ...newProps }: IComboBoxOptionWrapperProps,
+  ) =>
     // The render function will always be different, so we ignore that prop
-    return !shallowCompare({ ...this.props, render: undefined }, { ...newProps, render: undefined });
-  }
-}
+    shallowCompare(oldProps, newProps),
+);
 
 const COMPONENT_NAME = 'ComboBox';
+const DEFAULT_PROPS: Partial<IComboBoxProps> = {
+  options: [],
+  allowFreeform: false,
+  autoComplete: 'on',
+  buttonIconProps: { iconName: 'ChevronDown' },
+};
+
+function useOptionsState({ options, defaultSelectedKey, selectedKey }: IComboBoxProps) {
+  /** The currently selected indices */
+  const [selectedIndices, setSelectedIndices] = React.useState<number[]>(() =>
+    getSelectedIndices(options, buildDefaultSelectedKeys(defaultSelectedKey, selectedKey)),
+  );
+  /** The options currently available for the callout */
+  const [currentOptions, setCurrentOptions] = React.useState<IComboBoxOption[]>(options);
+  /** This value is used for the autocomplete hint value */
+  const [suggestedDisplayValue, setSuggestedDisplayValue] = React.useState<string>();
+
+  React.useEffect(() => {
+    if (selectedKey !== undefined) {
+      const selectedKeys: string[] | number[] = buildSelectedKeys(selectedKey);
+      const indices: number[] = getSelectedIndices(options, selectedKeys);
+
+      setSelectedIndices(indices);
+    }
+    setCurrentOptions(options);
+  }, [options, selectedKey]);
+
+  React.useEffect(() => {
+    if (selectedKey === null) {
+      setSuggestedDisplayValue(undefined);
+    }
+  }, [selectedKey]);
+
+  return [
+    selectedIndices,
+    setSelectedIndices,
+    currentOptions,
+    setCurrentOptions,
+    suggestedDisplayValue,
+    setSuggestedDisplayValue,
+  ] as const;
+}
+
+export const ComboBox: React.FunctionComponent<IComboBoxProps> = React.forwardRef<HTMLDivElement, IComboBoxProps>(
+  (propsWithoutDefaults: IComboBoxProps, forwardedRef: React.Ref<HTMLDivElement>) => {
+    const { ref, ...props } = getPropsWithDefaults(DEFAULT_PROPS, propsWithoutDefaults);
+    const rootRef = React.useRef<HTMLDivElement>(null);
+
+    const mergedRootRef = useMergedRefs(rootRef, forwardedRef);
+
+    const [
+      selectedIndices,
+      setSelectedIndices,
+      currentOptions,
+      setCurrentOptions,
+      suggestedDisplayValue,
+      setSuggestedDisplayValue,
+    ] = useOptionsState(props);
+
+    return (
+      <ComboBoxInternal
+        {...props}
+        hoisted={{
+          mergedRootRef,
+          rootRef,
+          selectedIndices,
+          setSelectedIndices,
+          currentOptions,
+          setCurrentOptions,
+          suggestedDisplayValue,
+          setSuggestedDisplayValue,
+        }}
+      />
+    );
+  },
+);
+ComboBox.displayName = COMPONENT_NAME;
+
+interface IComboBoxInternalProps extends Omit<IComboBoxProps, 'ref'> {
+  hoisted: {
+    mergedRootRef: React.Ref<HTMLDivElement>;
+    rootRef: React.RefObject<HTMLDivElement>;
+    selectedIndices: number[];
+    currentOptions: IComboBoxOption[];
+    suggestedDisplayValue?: string;
+    setSelectedIndices: React.Dispatch<React.SetStateAction<number[]>>;
+    setCurrentOptions: React.Dispatch<React.SetStateAction<IComboBoxOption[]>>;
+    setSuggestedDisplayValue: React.Dispatch<React.SetStateAction<string | undefined>>;
+  };
+}
 
 @customizable('ComboBox', ['theme', 'styles'], true)
-export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
-  public static defaultProps: IComboBoxProps = {
-    options: [],
-    allowFreeform: false,
-    autoComplete: 'on',
-    buttonIconProps: { iconName: 'ChevronDown' },
-  };
-
-  private _root = React.createRef<HTMLDivElement>();
-
-  /** The input aspect of the comboBox */
+class ComboBoxInternal extends React.Component<IComboBoxInternalProps, IComboBoxState> {
+  /** The input aspect of the combo box */
   private _autofill = React.createRef<IAutofill>();
 
   /** The wrapping div of the input and button */
@@ -186,7 +251,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _async: Async;
   private _events: EventGroup;
 
-  constructor(props: IComboBoxProps) {
+  constructor(props: IComboBoxInternalProps) {
     super(props);
 
     initializeComponentRef(this);
@@ -201,24 +266,15 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
     });
 
     this._id = props.id || getId('ComboBox');
-    const selectedKeys: string[] | number[] = this._buildDefaultSelectedKeys(
-      props.defaultSelectedKey,
-      props.selectedKey,
-    );
 
     this._isScrollIdle = true;
     this._processingTouch = false;
     this._gotMouseMove = false;
     this._processingClearPendingInfo = false;
 
-    const initialSelectedIndices: number[] = this._getSelectedIndices(props.options, selectedKeys);
-
     this.state = {
       isOpen: false,
-      selectedIndices: initialSelectedIndices,
       focusState: 'none',
-      suggestedDisplayValue: undefined,
-      currentOptions: this.props.options,
       currentPendingValueValidIndex: -1,
       currentPendingValue: undefined,
       currentPendingValueValidIndexOnHover: HoverStatus.default,
@@ -229,7 +285,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * All selected options
    */
   public get selectedOptions(): IComboBoxOption[] {
-    const { currentOptions, selectedIndices } = this.state;
+    const { currentOptions, selectedIndices } = this.props.hoisted;
 
     return getAllSelectedOptions(currentOptions, selectedIndices!);
   }
@@ -247,32 +303,15 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
     }
   }
 
-  public UNSAFE_componentWillReceiveProps(newProps: IComboBoxProps): void {
-    // Update the selectedIndex and currentOptions state if
-    // the selectedKey, value, or options have changed
-    if (
-      newProps.selectedKey !== this.props.selectedKey ||
-      newProps.text !== this.props.text ||
-      newProps.options !== this.props.options
-    ) {
-      const selectedKeys: string[] | number[] = this._buildSelectedKeys(newProps.selectedKey);
-      const indices: number[] = this._getSelectedIndices(newProps.options, selectedKeys);
-
-      this.setState({
-        selectedIndices: indices,
-        currentOptions: newProps.options,
-      });
-      if (newProps.selectedKey === null) {
-        this.setState({
-          suggestedDisplayValue: undefined,
-        });
-      }
-    }
-  }
-
-  public componentDidUpdate(prevProps: IComboBoxProps, prevState: IComboBoxState) {
-    const { allowFreeform, text, onMenuOpen, onMenuDismissed } = this.props;
-    const { isOpen, selectedIndices, currentPendingValueValidIndex } = this.state;
+  public componentDidUpdate(prevProps: IComboBoxInternalProps, prevState: IComboBoxState) {
+    const {
+      allowFreeform,
+      text,
+      onMenuOpen,
+      onMenuDismissed,
+      hoisted: { selectedIndices },
+    } = this.props;
+    const { isOpen, currentPendingValueValidIndex } = this.state;
 
     // If we are newly open or are open and the pending valid index changed,
     // make sure the currently selected/pending option is scrolled into view
@@ -309,9 +348,9 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         (this._hasFocus() &&
           ((!isOpen &&
             !this.props.multiSelect &&
-            prevState.selectedIndices &&
+            prevProps.hoisted.selectedIndices &&
             selectedIndices &&
-            prevState.selectedIndices[0] !== selectedIndices[0]) ||
+            prevProps.hoisted.selectedIndices[0] !== selectedIndices[0]) ||
             !allowFreeform ||
             text !== prevProps.text)))
     ) {
@@ -354,15 +393,16 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       keytipProps,
       persistMenu,
       multiSelect,
+      hoisted: { suggestedDisplayValue, selectedIndices, currentOptions },
     } = this.props;
-    const { isOpen, suggestedDisplayValue } = this.state;
+    const { isOpen } = this.state;
     this._currentVisibleValue = this._getVisibleValue();
 
     // Single select is already accessible since the whole text is selected
     // when focus enters the input. Since multiselect appears to clear the input
     // it needs special accessible text
     const multiselectAccessibleText = multiSelect
-      ? this._getMultiselectDisplayString(this.state.selectedIndices, this.state.currentOptions, suggestedDisplayValue)
+      ? this._getMultiselectDisplayString(selectedIndices, currentOptions, suggestedDisplayValue)
       : undefined;
 
     const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties, [
@@ -396,16 +436,16 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
 
     const comboBoxWrapper = keytipProps ? (
       <KeytipData keytipProps={keytipProps} disabled={disabled}>
-        {(keytipAttributes: any): JSX.Element =>
-          this._renderComboBoxWrapper(multiselectAccessibleText, errorMessageId, keytipAttributes)
-        }
+        {// eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (keytipAttributes: any): JSX.Element =>
+          this._renderComboBoxWrapper(multiselectAccessibleText, errorMessageId, keytipAttributes)}
       </KeytipData>
     ) : (
       this._renderComboBoxWrapper(multiselectAccessibleText, errorMessageId)
     );
 
     return (
-      <div {...divProps} ref={this._root} className={this._classNames.container}>
+      <div {...divProps} ref={this.props.hoisted.mergedRootRef} className={this._classNames.container}>
         {onRenderLabel({ props: this.props, multiselectAccessibleText }, this._onRenderLabel)}
         {comboBoxWrapper}
         {(persistMenu || isOpen) &&
@@ -415,7 +455,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
               onRenderList,
               onRenderItem,
               onRenderOption,
-              options: this.state.currentOptions.map((item, index) => ({ ...item, index: index })),
+              options: currentOptions.map((item, index) => ({ ...item, index: index })),
               onDismiss: this._onDismiss,
             },
             this._onRenderContainer,
@@ -451,7 +491,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       }
     }
 
-    // Programatically setting focus means that there is nothing else that needs to be done
+    // Programmatically setting focus means that there is nothing else that needs to be done
     // Focus is now contained
     if (!this._hasFocus()) {
       this.setState({ focusState: 'focused' });
@@ -468,7 +508,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
 
   /**
    * componentWillReceiveProps handler for the auto fill component
-   * Checks/updates the iput value to set, if needed
+   * Checks/updates the input value to set, if needed
    * @param defaultVisibleValue - the defaultVisibleValue that got passed
    *  in to the auto fill's componentWillReceiveProps
    * @returns - the updated value to set, if needed
@@ -496,6 +536,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _renderComboBoxWrapper = (
     multiselectAccessibleText: string | undefined,
     errorMessageId: string,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     keytipAttributes: any = {},
   ): JSX.Element => {
     const {
@@ -513,14 +554,15 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       tabIndex,
       autofill,
       iconButtonProps,
+      hoisted: { suggestedDisplayValue },
     } = this.props;
 
-    const { isOpen, suggestedDisplayValue } = this.state;
+    const { isOpen } = this.state;
 
-    // If the combobox has focus, is multiselect, and has a display string, then use that placeholder
+    // If the combo box has focus, is multiselect, and has a display string, then use that placeholder
     // so that the selected items don't appear to vanish. This is not ideal but it's the only reasonable way
     // to correct the behavior where the input is cleared so the user can type. If a full refactor is done, then this
-    // should be removed and the multiselect combobox should behave like a picker.
+    // should be removed and the multiselect combo box should behave like a picker.
     const placeholder =
       this._hasFocus() && this.props.multiSelect && multiselectAccessibleText
         ? multiselectAccessibleText
@@ -602,7 +644,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * True if the defaultVisibleValue equals the suggestedDisplayValue, false otherwise
    */
   private _onShouldSelectFullInputValueInAutofillComponentDidUpdate = (): boolean => {
-    return this._currentVisibleValue === this.state.suggestedDisplayValue;
+    return this._currentVisibleValue === this.props.hoisted.suggestedDisplayValue;
   };
 
   /**
@@ -611,15 +653,13 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * @returns the value to pass to the input
    */
   private _getVisibleValue = (): string | undefined => {
-    const { text, allowFreeform, autoComplete } = this.props;
     const {
-      selectedIndices,
-      currentPendingValueValidIndex,
-      currentOptions,
-      currentPendingValue,
-      suggestedDisplayValue,
-      isOpen,
-    } = this.state;
+      text,
+      allowFreeform,
+      autoComplete,
+      hoisted: { suggestedDisplayValue, selectedIndices, currentOptions },
+    } = this.props;
+    const { currentPendingValueValidIndex, currentPendingValue, isOpen } = this.state;
 
     const currentPendingIndexValid = this._indexWithinBounds(currentOptions, currentPendingValueValidIndex);
 
@@ -695,7 +735,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
 
   /**
    * Returns a string that concatenates all of the selected values
-   * for multiselect combobox.
+   * for multiselect combo box.
    */
   private _getMultiselectDisplayString(
     selectedIndices: number[] | undefined,
@@ -744,12 +784,12 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   };
 
   /**
-   * Process the new input's new value when the comboBox
+   * Process the new input's new value when the combo box
    * allows freeform entry
    * @param updatedValue - the input's newly changed value
    */
   private _processInputChangeWithFreeform(updatedValue: string): void {
-    const { currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
     let newCurrentPendingValueValidIndex = -1;
 
     // if the new value is empty, see if we have an exact match
@@ -834,24 +874,25 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Process the new input's new value when the comboBox
+   * Process the new input's new value when the combo box
    * does not allow freeform entry
    * @param updatedValue - the input's newly changed value
    */
   private _processInputChangeWithoutFreeform(updatedValue: string): void {
-    const { currentPendingValue, currentPendingValueValidIndex, currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
+    const { currentPendingValue, currentPendingValueValidIndex } = this.state;
 
     if (this.props.autoComplete === 'on') {
       // If autoComplete is on while allow freeform is off,
-      // we will remember the keypresses and build up a string to attempt to match
+      // we will remember the key press and build up a string to attempt to match
       // as long as characters are typed within a the timeout span of each other,
       // otherwise we will clear the string and start building a new one on the next keypress.
       // Also, only do this processing if we have a non-empty value
       if (updatedValue !== '') {
         // If we have a pending autocomplete clearing task,
-        // we know that the user is typing with keypresses happening
+        // we know that the user is typing with key press happening
         // within the timeout of each other so remove the clearing task
-        // and continue building the pending value with the udpated value
+        // and continue building the pending value with the updated value
         if (this._lastReadOnlyAutoCompleteChangeTimeoutId !== undefined) {
           this._async.clearTimeout(this._lastReadOnlyAutoCompleteChangeTimeoutId);
           this._lastReadOnlyAutoCompleteChangeTimeoutId = undefined;
@@ -873,7 +914,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
           )
           .filter(option => option.text.toLocaleLowerCase().indexOf(updatedValue) === 0);
 
-        // If we found a match, udpdate the state
+        // If we found a match, update the state
         if (items.length > 0) {
           this._setPendingInfo(originalUpdatedValue, items[0].index, this._getPreviewText(items[0]));
         }
@@ -899,7 +940,9 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   private _getFirstSelectedIndex(): number {
-    return this.state.selectedIndices && this.state.selectedIndices.length > 0 ? this.state.selectedIndices[0] : -1;
+    return this.props.hoisted.selectedIndices && this.props.hoisted.selectedIndices.length > 0
+      ? this.props.hoisted.selectedIndices[0]
+      : -1;
   }
 
   /**
@@ -911,7 +954,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * it will snap to the edge of the options array. If delta == 0 and the given index is not selectable
    */
   private _getNextSelectableIndex(index: number, searchDirection: SearchDirection): number {
-    const { currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
 
     let newIndex = index + searchDirection;
 
@@ -953,12 +996,15 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    */
   private _setSelectedIndex(
     index: number,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     submitPendingValueEvent: React.SyntheticEvent<any>,
     searchDirection: SearchDirection = SearchDirection.none,
   ): void {
-    const { onChange, onPendingValueChanged } = this.props;
-    const { currentOptions } = this.state;
-    const { selectedIndices: initialIndices } = this.state;
+    const {
+      onChange,
+      onPendingValueChanged,
+      hoisted: { selectedIndices: initialIndices, currentOptions },
+    } = this.props;
 
     // Clone selectedIndices so we don't mutate state
     let selectedIndices = initialIndices ? initialIndices.slice() : [];
@@ -984,7 +1030,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         return;
       }
       if (this.props.multiSelect) {
-        // Setting the initial state of option.selected in Multi-select combobox by checking the
+        // Setting the initial state of option.selected in Multi-select combo box by checking the
         // selectedIndices array and overriding the undefined issue
         option.selected = option.selected !== undefined ? !option.selected : selectedIndices.indexOf(index) < 0;
         if (option.selected && selectedIndices.indexOf(index) < 0) {
@@ -998,9 +1044,9 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
 
       submitPendingValueEvent.persist();
 
-      // Only setstate if combobox is uncontrolled.
+      // Only setState if combo box is uncontrolled.
       if (this.props.selectedKey || this.props.selectedKey === null) {
-        // If ComboBox value is changed, revert preview first
+        // If combo box value is changed, revert preview first
         if (this._hasPendingValue && onPendingValueChanged) {
           onPendingValueChanged();
           this._hasPendingValue = false;
@@ -1014,23 +1060,18 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         changedOptions[index] = option;
 
         // Call onChange after state is updated
-        this.setState(
-          {
-            selectedIndices: selectedIndices,
-            currentOptions: changedOptions,
-          },
-          () => {
-            // If ComboBox value is changed, revert preview first
-            if (this._hasPendingValue && onPendingValueChanged) {
-              onPendingValueChanged();
-              this._hasPendingValue = false;
-            }
+        this.props.hoisted.setSelectedIndices(selectedIndices);
+        this.props.hoisted.setCurrentOptions(changedOptions);
 
-            if (onChange) {
-              onChange(submitPendingValueEvent, option, index, undefined);
-            }
-          },
-        );
+        // If ComboBox value is changed, revert preview first
+        if (this._hasPendingValue && onPendingValueChanged) {
+          onPendingValueChanged();
+          this._hasPendingValue = false;
+        }
+
+        if (onChange) {
+          onChange(submitPendingValueEvent, option, index, undefined);
+        }
       }
     }
     if (this.props.multiSelect && this.state.isOpen) {
@@ -1062,24 +1103,20 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _onResolveOptions = (): void => {
     if (this.props.onResolveOptions) {
       // get the options
-      const newOptions = this.props.onResolveOptions([...this.state.currentOptions]);
+      const newOptions = this.props.onResolveOptions([...this.props.hoisted.currentOptions]);
 
       // Check to see if the returned value is an array, if it is update the state
       // If the returned value is not an array then check to see if it's a promise or PromiseLike.
       // If it is then resolve it asynchronously.
       if (Array.isArray(newOptions)) {
-        this.setState({
-          currentOptions: newOptions,
-        });
+        this.props.hoisted.setCurrentOptions(newOptions);
       } else if (newOptions && newOptions.then) {
         // Ensure that the promise will only use the callback if it was the most recent one
         // and update the state when the promise returns
         const promise: PromiseLike<IComboBoxOption[]> = (this._currentPromise = newOptions);
         promise.then((newOptionsFromPromise: IComboBoxOption[]) => {
           if (promise === this._currentPromise) {
-            this.setState({
-              currentOptions: newOptionsFromPromise,
-            });
+            this.props.hoisted.setCurrentOptions(newOptionsFromPromise);
           }
         });
       }
@@ -1094,7 +1131,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _onBlur = (event: React.FocusEvent<HTMLElement | Autofill | BaseButton | Button>): void => {
     // Do nothing if the blur is coming from something
     // inside the comboBox root or the comboBox menu since
-    // it we are not really bluring from the whole comboBox
+    // it we are not really blurring from the whole comboBox
     let relatedTarget = event.relatedTarget;
     if (event.relatedTarget === null) {
       // In IE11, due to lack of support, event.relatedTarget is always
@@ -1107,7 +1144,8 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
     if (
       relatedTarget &&
       // when event coming from withing the comboBox title
-      ((this._root.current && this._root.current.contains(relatedTarget as HTMLElement)) ||
+      ((this.props.hoisted.rootRef.current &&
+        this.props.hoisted.rootRef.current.contains(relatedTarget as HTMLElement)) ||
         // when event coming from within the comboBox list menu
         (this._comboBoxMenu.current &&
           (this._comboBoxMenu.current.contains(relatedTarget as HTMLElement) ||
@@ -1131,15 +1169,16 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   /**
    * Submit a pending value if there is one
    */
-  private _submitPendingValue(submitPendingValueEvent: any): void {
-    const { onChange, allowFreeform, autoComplete } = this.props;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  private _submitPendingValue(submitPendingValueEvent: React.SyntheticEvent<any>): void {
     const {
-      currentPendingValue,
-      currentPendingValueValidIndex,
-      currentOptions,
-      currentPendingValueValidIndexOnHover,
-    } = this.state;
-    let { selectedIndices } = this.state;
+      onChange,
+      allowFreeform,
+      autoComplete,
+      hoisted: { currentOptions },
+    } = this.props;
+    const { currentPendingValue, currentPendingValueValidIndex, currentPendingValueValidIndexOnHover } = this.state;
+    let { selectedIndices } = this.props.hoisted;
 
     // Do not submit any pending value if we
     // have already initiated clearing the pending info
@@ -1215,10 +1254,8 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
           }
           selectedIndices.push(newOptions.length - 1);
         }
-        this.setState({
-          currentOptions: newOptions,
-          selectedIndices: selectedIndices,
-        });
+        this.props.hoisted.setCurrentOptions(newOptions);
+        this.props.hoisted.setSelectedIndices(selectedIndices);
       }
     } else if (currentPendingValueValidIndex >= 0) {
       // Since we are not allowing freeform, we must have a matching
@@ -1280,7 +1317,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       >
         {onRenderUpperContent(this.props, this._onRenderUpperContent)}
         <div className={this._classNames.optionsContainerWrapper} ref={this._comboBoxMenu}>
-          {(onRenderList as any)({ ...props }, this._onRenderList)}
+          {onRenderList?.({ ...props }, this._onRenderList)}
         </div>
         {onRenderLowerContent(this.props, this._onRenderLowerContent)}
       </Callout>
@@ -1332,7 +1369,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         aria-labelledby={id + '-label'}
         role="listbox"
       >
-        {options.map(item => (onRenderItem as any)(item, this._onRenderItem))}
+        {options.map(item => onRenderItem?.(item, this._onRenderItem))}
       </div>
     );
   };
@@ -1481,10 +1518,10 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   private _isOptionChecked(index: number | undefined): boolean {
-    if (this.props.multiSelect && index !== undefined && this.state.selectedIndices) {
+    if (this.props.multiSelect && index !== undefined && this.props.hoisted.selectedIndices) {
       let idxOfSelectedIndex = -1;
 
-      idxOfSelectedIndex = this.state.selectedIndices.indexOf(index);
+      idxOfSelectedIndex = this.props.hoisted.selectedIndices.indexOf(index);
       return idxOfSelectedIndex >= 0;
     }
     return false;
@@ -1591,11 +1628,12 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * to select the item and also close the menu
    * @param index - the index of the item that was clicked
    */
-  private _onItemClick(item: IComboBoxOption): (ev: any) => void {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  private _onItemClick(item: IComboBoxOption): (ev: React.MouseEvent<any>) => void {
     const { onItemClick } = this.props;
     const { index } = item;
-
-    return (ev: any): void => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return (ev: React.MouseEvent<any>): void => {
       onItemClick && onItemClick(ev, item, index);
       this._setSelectedIndex(index as number, ev);
 
@@ -1635,58 +1673,21 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   };
 
   /**
-   * Get the indices of the options that are marked as selected
-   * @param options - the comboBox options
-   * @param selectedKeys - the known selected keys to find
-   * @returns - an array of the indices of the selected options, empty array if nothing is selected
-   */
-  private _getSelectedIndices(
-    options: IComboBoxOption[] | undefined,
-    selectedKeys: (string | number | undefined)[],
-  ): number[] {
-    if (!options || !selectedKeys) {
-      return [];
-    }
-
-    const selectedIndices: { [key: number]: boolean } = {};
-    options.forEach((option: IComboBoxOption, index: number) => {
-      if (option.selected) {
-        selectedIndices[index] = true;
-      }
-    });
-
-    for (const selectedKey of selectedKeys) {
-      const index = findIndex(options, option => option.key === selectedKey);
-      if (index > -1) {
-        selectedIndices[index] = true;
-      }
-    }
-
-    return Object.keys(selectedIndices)
-      .map(Number)
-      .sort();
-  }
-
-  /**
    * Reset the selected index by clearing the
    * input (of any pending text), clearing the pending state,
    * and setting the suggested display value to the last
    * selected state text
    */
   private _resetSelectedIndex(): void {
-    const { currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
     this._clearPendingInfo();
 
     const selectedIndex: number = this._getFirstSelectedIndex();
     if (selectedIndex > 0 && selectedIndex < currentOptions.length) {
-      this.setState({
-        suggestedDisplayValue: currentOptions[selectedIndex].text,
-      });
+      this.props.hoisted.setSuggestedDisplayValue(currentOptions[selectedIndex].text);
     } else if (this.props.text) {
       // If we had a value initially, restore it
-      this.setState({
-        suggestedDisplayValue: this.props.text,
-      });
+      this.props.hoisted.setSuggestedDisplayValue(this.props.text);
     }
   }
 
@@ -1696,11 +1697,11 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _clearPendingInfo(): void {
     this._processingClearPendingInfo = true;
 
+    this.props.hoisted.setSuggestedDisplayValue(undefined);
     this.setState(
       {
         currentPendingValue: undefined,
         currentPendingValueValidIndex: -1,
-        suggestedDisplayValue: undefined,
         currentPendingValueValidIndexOnHover: HoverStatus.default,
       },
       this._onAfterClearPendingInfo,
@@ -1726,10 +1727,10 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       return;
     }
 
+    this.props.hoisted.setSuggestedDisplayValue(suggestedDisplayValue);
     this.setState({
       currentPendingValue: this._normalizeToString(currentPendingValue),
       currentPendingValueValidIndex: currentPendingValueValidIndex,
-      suggestedDisplayValue: suggestedDisplayValue,
       currentPendingValueValidIndexOnHover: HoverStatus.default,
     });
   }
@@ -1739,7 +1740,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * @param index - the index to set the pending info from
    */
   private _setPendingInfoFromIndex(index: number): void {
-    const { currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
 
     if (index >= 0 && index < currentOptions.length) {
       const option = currentOptions[index];
@@ -1750,12 +1751,12 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Sets the pending info for the comboBox
+   * Sets the pending info for the combo box
    * @param index - the index to search from
    * @param searchDirection - the direction to search
    */
   private _setPendingInfoFromIndexAndDirection(index: number, searchDirection: SearchDirection): void {
-    const { currentOptions } = this.state;
+    const { currentOptions } = this.props.hoisted;
 
     // update index to allow content to wrap
     if (searchDirection === SearchDirection.forward && index >= currentOptions.length - 1) {
@@ -1767,7 +1768,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
     // get the next "valid" index
     const indexUpdate = this._getNextSelectableIndex(index, searchDirection);
 
-    // if the two indicies are equal we didn't move and
+    // if the two indices are equal we didn't move and
     // we should attempt to get  get the first/last "valid" index to use
     // (Note, this takes care of the potential cases where the first/last
     // item is not focusable), otherwise use the updated index
@@ -1793,12 +1794,8 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
       return;
     }
 
-    const {
-      currentPendingValue,
-      currentOptions,
-      currentPendingValueValidIndex,
-      currentPendingValueValidIndexOnHover,
-    } = this.state;
+    const { currentOptions } = this.props.hoisted;
+    const { currentPendingValue, currentPendingValueValidIndex, currentPendingValueValidIndexOnHover } = this.state;
 
     let newPendingIndex: number | undefined = undefined;
     let newPendingValue: string | undefined = undefined;
@@ -1846,8 +1843,13 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    * @param ev - The keyboard event that was fired
    */
   private _onInputKeyDown = (ev: React.KeyboardEvent<HTMLElement | Autofill>): void => {
-    const { disabled, allowFreeform, autoComplete } = this.props;
-    const { isOpen, currentOptions, currentPendingValueValidIndexOnHover } = this.state;
+    const {
+      disabled,
+      allowFreeform,
+      autoComplete,
+      hoisted: { currentOptions },
+    } = this.props;
+    const { isOpen, currentPendingValueValidIndexOnHover } = this.state;
 
     // Take note if we are processing an alt (option) or meta (command) keydown.
     // See comment in _onInputKeyUp for reasoning.
@@ -1900,12 +1902,12 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         }
 
         // If we are not allowing freeform
-        // or the comboBox is open, flip the open state
+        // or the combo box is open, flip the open state
         if (isOpen) {
           this._setOpenStateAndFocusOnClose(!isOpen, false /* focusInputAfterClose */);
         }
 
-        // Allow TAB to propigate
+        // Allow TAB to propagate
         return;
 
       case KeyCodes.escape:
@@ -1927,7 +1929,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         // and has since mousedOut of the menu items),
         // go to the last index
         if (currentPendingValueValidIndexOnHover === HoverStatus.clearAll) {
-          index = this.state.currentOptions.length;
+          index = this.props.hoisted.currentOptions.length;
         }
 
         if (ev.altKey || ev.metaKey) {
@@ -1947,7 +1949,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
         break;
 
       case KeyCodes.down:
-        // Expand the comboBox on ALT + DownArrow
+        // Expand the combo box on ALT + DownArrow
         if (ev.altKey || ev.metaKey) {
           this._setOpenStateAndFocusOnClose(true /* isOpen */, true /* focusInputAfterClose */);
         } else {
@@ -2054,7 +2056,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
     switch (ev.which) {
       case KeyCodes.space:
         // If we are not allowing freeform and are not autoComplete
-        // make space expand/collapse the comboBox
+        // make space expand/collapse the combo box
         // and allow the event to propagate
         if (!allowFreeform && autoComplete === 'off') {
           this._setOpenStateAndFocusOnClose(!isOpen, !!isOpen);
@@ -2124,7 +2126,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    */
   private _handleInputWhenDisabled(ev: React.KeyboardEvent<HTMLElement | Autofill> | null): void {
     // If we are disabled, close the menu (if needed)
-    // and eat all keystokes other than TAB or ESC
+    // and eat all keystrokes other than TAB or ESC
     if (this.props.disabled) {
       if (this.state.isOpen) {
         this.setState({ isOpen: false });
@@ -2145,9 +2147,9 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Click handler for the button of the comboBox
+   * Click handler for the button of the combo box
    * and the input when not allowing freeform. This
-   * toggles the expand/collapse state of the comboBox (if enbled)
+   * toggles the expand/collapse state of the combo box (if enabled)
    */
   private _onComboBoxClick = (): void => {
     const { disabled } = this.props;
@@ -2164,7 +2166,6 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
    */
   private _onAutofillClick = (): void => {
     const { disabled, allowFreeform } = this.props;
-
     if (allowFreeform && !disabled) {
       this.focus(this.state.isOpen || this._processingTouch);
     } else {
@@ -2188,7 +2189,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   };
 
   private _handleTouchAndPointerEvent() {
-    // If we already have an existing timeeout from a previous touch and pointer event
+    // If we already have an existing timeout from a previous touch and pointer event
     // cancel that timeout so we can set a nwe one.
     if (this._lastTouchTimeoutId !== undefined) {
       this._async.clearTimeout(this._lastTouchTimeoutId);
@@ -2229,14 +2230,14 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Get the aria-activedescendant value for the comboxbox.
+   * Get the aria-activedescendant value for the combo box.
    * @returns the id of the current focused combo item, otherwise the id of the currently selected element,
    * null otherwise
    */
   private _getAriaActiveDescendantValue(): string | undefined {
     let descendantText =
-      this.state.isOpen && this.state.selectedIndices && this.state.selectedIndices.length > 0
-        ? this._id + '-list' + this.state.selectedIndices[0]
+      this.state.isOpen && this.props.hoisted.selectedIndices && this.props.hoisted.selectedIndices.length > 0
+        ? this._id + '-list' + this.props.hoisted.selectedIndices[0]
         : undefined;
     if (this.state.isOpen && this._hasFocus() && this.state.currentPendingValueValidIndex !== -1) {
       descendantText = this._id + '-list' + this.state.currentPendingValueValidIndex;
@@ -2245,7 +2246,7 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   }
 
   /**
-   * Get the aria autocomplete value for the Combobox
+   * Get the aria autocomplete value for the combo box
    * @returns 'inline' if auto-complete automatically dynamic, 'both' if we have a list of possible values to pick from
    * and can dynamically populate input, and 'none' if auto-complete is not enabled as we can't give user inputs.
    */
@@ -2256,35 +2257,6 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
 
   private _isPendingOption(item: IComboBoxOption): boolean {
     return item && item.index === this.state.currentPendingValueValidIndex;
-  }
-
-  /**
-   * Given default selected key(s) and selected key(s), return the selected keys(s).
-   * When default selected key(s) are available, they take precedence and return them instead of selected key(s).
-   *
-   * @returns No matter what specific types the input parameters are, always return an array of
-   *  either strings or numbers instead of premitive type.  This normlization makes caller's logic easier.
-   */
-  private _buildDefaultSelectedKeys(
-    defaultSelectedKey: string | number | string[] | number[] | null | undefined,
-    selectedKey: string | number | string[] | number[] | null | undefined,
-  ): string[] | number[] {
-    const selectedKeys: string[] | number[] = this._buildSelectedKeys(defaultSelectedKey);
-    if (selectedKeys.length) {
-      return selectedKeys;
-    }
-    return this._buildSelectedKeys(selectedKey);
-  }
-
-  private _buildSelectedKeys(
-    selectedKey: string | number | string[] | number[] | null | undefined,
-  ): string[] | number[] {
-    if (selectedKey === undefined) {
-      return [];
-    }
-
-    // need to cast here so typescript does not complain
-    return (selectedKey instanceof Array ? selectedKey : [selectedKey]) as string[] | number[];
   }
 
   // For scenarios where the option's text prop contains embedded styles, we use the option's
@@ -2304,4 +2276,64 @@ export class ComboBox extends React.Component<IComboBoxProps, IComboBoxState> {
   private _hasFocus() {
     return this.state.focusState !== 'none';
   }
+}
+
+/**
+ * Get the indices of the options that are marked as selected
+ * @param options - the combo box options
+ * @param selectedKeys - the known selected keys to find
+ * @returns - an array of the indices of the selected options, empty array if nothing is selected
+ */
+function getSelectedIndices(
+  options: IComboBoxOption[] | undefined,
+  selectedKeys: (string | number | undefined)[],
+): number[] {
+  if (!options || !selectedKeys) {
+    return [];
+  }
+
+  const selectedIndices: { [key: number]: boolean } = {};
+  options.forEach((option: IComboBoxOption, index: number) => {
+    if (option.selected) {
+      selectedIndices[index] = true;
+    }
+  });
+
+  for (const selectedKey of selectedKeys) {
+    const index = findIndex(options, option => option.key === selectedKey);
+    if (index > -1) {
+      selectedIndices[index] = true;
+    }
+  }
+
+  return Object.keys(selectedIndices)
+    .map(Number)
+    .sort();
+}
+
+/**
+ * Given default selected key(s) and selected key(s), return the selected keys(s).
+ * When default selected key(s) are available, they take precedence and return them instead of selected key(s).
+ *
+ * @returns No matter what specific types the input parameters are, always return an array of
+ *  either strings or numbers instead of primitive type.  This normalization makes caller's logic easier.
+ */
+function buildDefaultSelectedKeys(
+  defaultSelectedKey: string | number | string[] | number[] | null | undefined,
+  selectedKey: string | number | string[] | number[] | null | undefined,
+): string[] | number[] {
+  const selectedKeys: string[] | number[] = buildSelectedKeys(defaultSelectedKey);
+  if (selectedKeys.length) {
+    return selectedKeys;
+  }
+  return buildSelectedKeys(selectedKey);
+}
+
+function buildSelectedKeys(selectedKey: string | number | string[] | number[] | null | undefined): string[] | number[] {
+  if (selectedKey === undefined) {
+    return [];
+  }
+
+  // need to cast here so typescript does not complain
+  return (selectedKey instanceof Array ? selectedKey : [selectedKey]) as string[] | number[];
 }
