+++ packages/react/src/components/ComboBox/ComboBox.types.ts
@@ -2,30 +2,30 @@ import * as React from 'react';
 import { IIconProps } from '../../Icon';
 import { ISelectableOption, ISelectableDroppableTextProps } from '../../SelectableOption';
 import { IStyle, ITheme } from '../../Styling';
-import { IButtonStyles } from '../../Button';
+import { IButtonStyles, IButtonProps } from '../../compat/Button';
 import { IRefObject, IRenderFunction } from '../../Utilities';
+import { IComboBoxClassNames } from './ComboBox.classNames';
 import { IKeytipProps } from '../../Keytip';
 import { IAutofillProps } from '../../Autofill';
-import { IButtonProps } from '@fluentui/react-internal/lib/Button';
 
 /**
  * {@docCategory ComboBox}
  */
 export interface IComboBox {
   /**
-   * All selected options
+   * All selected options.
    */
   readonly selectedOptions: IComboBoxOption[];
 
   /**
-   * If there is a menu open this will dismiss the menu
+   * If there is a menu open this will dismiss the menu.
    */
   dismissMenu: () => void;
 
   /**
-   * Sets focus to the input in the comboBox
-   * @param shouldOpenOnFocus - Determines if we should open the ComboBox menu when the input gets focus
-   * @param useFocusAsync - Determines if we should focus the input asynchronously
+   * Sets focus to the input in the ComboBox
+   * @param shouldOpenOnFocus - Determines if we should open the ComboBox menu when the input gets focus.
+   * @param useFocusAsync - Determines if we should focus the input asynchronously.
    * @returns True if focus could be set, false if no operation was taken.
    */
   focus(shouldOpenOnFocus?: boolean, useFocusAsync?: boolean): boolean;
@@ -36,15 +36,15 @@ export interface IComboBox {
  */
 export interface IComboBoxOption extends ISelectableOption {
   /**
-   * Specific styles for each comboBox option. If you intend to give
-   * common styles to all comboBox option please use
-   * the prop comboBoxOptionStyles
+   * Specific styles for each ComboBox option. If you intend to give
+   * common styles to all ComboBox option please use
+   * the prop comboBoxOptionStyles.
    */
   styles?: Partial<IComboBoxOptionStyles>;
 
   /**
    * In scenarios where embedded data is used at the text prop, we will use the ariaLabel prop
-   * to set the aria-label and preview text. Default to false
+   * to set the aria-label and preview text. Default to false.
    * @defaultvalue false;
    */
   useAriaLabelAsText?: boolean;
@@ -53,7 +53,9 @@ export interface IComboBoxOption extends ISelectableOption {
 /**
  * {@docCategory ComboBox}
  */
-export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox, IComboBox> {
+export interface IComboBoxProps
+  extends ISelectableDroppableTextProps<IComboBox, IComboBox>,
+    React.RefAttributes<HTMLDivElement> {
   /**
    * Optional callback to access the IComboBox interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
@@ -61,7 +63,7 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   componentRef?: IRefObject<IComboBox>;
 
   /**
-   * Collection of options for this ComboBox
+   * Collection of options for this ComboBox.
    */
   options: IComboBoxOption[];
 
@@ -72,8 +74,8 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
 
   /**
    * Callback issued when either:
-   * 1) the selected option changes
-   * 2) a manually edited value is submitted. In this case there may not be a matched option if allowFreeform
+   * 1) The selected option changes.
+   * 2) A manually edited value is submitted. In this case there may not be a matched option if allowFreeform
    *    is also true (and hence only value would be true, the other parameter would be null in this case)
    */
   onChange?: (event: React.FormEvent<IComboBox>, option?: IComboBoxOption, index?: number, value?: string) => void;
@@ -87,39 +89,40 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   onPendingValueChanged?: (option?: IComboBoxOption, index?: number, value?: string) => void;
 
   /**
-   * Function that gets invoked when the ComboBox menu is launched
+   * Function that gets invoked when the ComboBox menu is launched.
    */
   onMenuOpen?: () => void;
 
   /**
-   * Function that gets invoked when the ComboBox menu is dismissed
+   * Function that gets invoked when the ComboBox menu is dismissed.
    */
   onMenuDismissed?: () => void;
 
   /**
-   * Function that gets invoked before the menu gets dismissed
+   * Function that gets invoked before the menu gets dismissed.
    */
   onMenuDismiss?: () => void;
 
   /**
    * Callback issued when the options should be resolved, if they have been updated or
-   * if they need to be passed in the first time
+   * if they need to be passed in the first time.
    */
   onResolveOptions?: (options: IComboBoxOption[]) => IComboBoxOption[] | PromiseLike<IComboBoxOption[]>;
 
   /**
-   * Callback issued when the ComboBox requests the list to scroll to a specific element
+   * Callback issued when the ComboBox requests the list to scroll to a specific element.
    */
   onScrollToItem?: (itemIndex: number) => void;
 
   /**
-   * Whether the ComboBox is free form, meaning that the user input is not bound to provided options. Defaults to false.
+   * Whether the ComboBox is free form, meaning that the user input is not bound to provided options.
+   * Defaults to false.
    */
   allowFreeform?: boolean;
 
   /**
-   * Whether the ComboBox auto completes. As the user is inputing text, it will be suggested potential matches from
-   * the list of options. If the combo box is expanded, this will also scroll to the suggested option, and give it a
+   * Whether the ComboBox auto completes. As the user is entering text, it will be suggested potential matches from
+   * the list of options. If the ComboBox is expanded, this will also scroll to the suggested option, and give it a
    * selected style.
    *
    * @defaultvalue "on"
@@ -127,24 +130,24 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   autoComplete?: 'on' | 'off';
 
   /**
-   * Value to show in the input, does not have to map to a combobox option
+   * Value to show in the input, does not have to map to a ComboBox option.
    */
   text?: string;
 
   /**
-   * When multiple items are selected, this will be used to separate values in the combobox input.
+   * When multiple items are selected, this will be used to separate values in the ComboBox input.
    *
    * @defaultvalue ", "
    */
   multiSelectDelimiter?: string;
 
   /**
-   * The IconProps to use for the button aspect of the combobox
+   * The IconProps to use for the button aspect of the ComboBox.
    */
   buttonIconProps?: IIconProps;
 
   /**
-   * The AutofillProps to be passed into the Autofill component inside combobox
+   * The AutofillProps to be passed into the Autofill component inside ComboBox.
    */
   autofill?: IAutofillProps;
 
@@ -154,7 +157,7 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   theme?: ITheme;
 
   /**
-   * Custom styles for this component
+   * Custom styles for this component.
    */
   styles?: Partial<IComboBoxStyles>;
 
@@ -207,12 +210,12 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   dropdownWidth?: number;
 
   /**
-   * Whether to use the ComboBoxes width as the menu's width
+   * Whether to use the ComboBox width as the menu's width.
    */
   useComboBoxAsMenuWidth?: boolean;
 
   /**
-   * Custom max width for dropdown
+   * Custom max width for dropdown.
    */
   dropdownMaxWidth?: number;
 
@@ -225,12 +228,12 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
   isButtonAriaHidden?: boolean;
 
   /**
-   * Optional prop to add a string id that can be referenced inside the aria-describedby attribute
+   * Optional prop to add a string id that can be referenced inside the aria-describedby attribute.
    */
   ariaDescribedBy?: string;
 
   /**
-   * Optional keytip for this combo box
+   * Optional keytip for this ComboBox.
    */
   keytipProps?: IKeytipProps;
 
@@ -239,20 +242,20 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
    * will be hidden. This will improve perf of the menu opening but could potentially
    * impact overall perf by having more elements in the dom. Should only be used
    * when perf is important.
-   * Note: This may increase the amount of time it takes for the comboBox itself to mount.
+   * Note: This may increase the amount of time it takes for the ComboBox itself to mount.
    */
   persistMenu?: boolean;
 
   /**
    * When specified, determines whether the callout (the menu which drops down) should
-   * restore the focus after being dismissed or not.  If false, then the menu will not try
+   * restore the focus after being dismissed or not. If false, then the menu will not try
    * to set focus to whichever element had focus before the menu was opened.
    * @defaultvalue true;
    */
   shouldRestoreFocus?: boolean;
 
   /**
-   * Optional iconButton props on combo box
+   * Optional iconButton props on ComboBox.
    */
   iconButtonProps?: IButtonProps;
 
@@ -267,12 +270,12 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox,
  */
 export interface IOnRenderComboBoxLabelProps {
   /**
-   * Props to render the combobox.
+   * Props to render the ComboBox.
    */
   props: IComboBoxProps;
 
   /**
-   * Accessible text for label when combobox is multiselected.
+   * Accessible text for label when ComboBox is multiselected.
    */
   multiselectAccessibleText?: string;
 }
@@ -282,7 +285,7 @@ export interface IOnRenderComboBoxLabelProps {
  */
 export interface IComboBoxStyles {
   /**
-   * Style for the container which has the ComboBox and the label
+   * Style for the container which has the ComboBox and the label.
    */
   container: IStyle;
 
@@ -312,42 +315,41 @@ export interface IComboBoxStyles {
   rootDisallowFreeForm: IStyle;
 
   /**
-   * Styles for when the ComboBox is hovered. These styles are applied for all comboBoxes except when
-   * the comboBox is disabled.
+   * Styles for when the ComboBox is hovered. These styles are applied for all ComboBoxes except when
+   * the ComboBox is disabled.
    */
   rootHovered: IStyle;
 
   /**
-   * Styles for when the ComboBox is active. These styles are applied for all comboBoxes except when
-   * the comboBox is disabled.
+   * Styles for when the ComboBox is active. These styles are applied for all ComboBoxes except when
+   * the ComboBox is disabled.
    */
   rootPressed: IStyle;
 
   /**
-   * Styles for when the ComboBox is focused. These styles are applied for all comboBoxes except when
-   * the comboBox is disabled.
+   * Styles for when the ComboBox is focused. These styles are applied for all ComboBoxes except when
+   * the ComboBox is disabled.
    */
   rootFocused: IStyle;
 
   /**
-   * Styles for when the comboBox is disabled. These styles override all the other styles.
-   * NOTE : Hover (or) Focused (or) active styles are not applied for disabled comboBoxes.
+   * Styles for when the ComboBox is disabled. These styles override all the other styles.
+   * NOTE : Hover (or) Focused (or) active styles are not applied for disabled ComboBoxes.
    */
   rootDisabled: IStyle;
 
   /**
-   * Base styles for the input element - which contains the currently selected
-   * option.
+   * Base styles for the input element - which contains the currently selected option.
    */
   input: IStyle;
 
   /**
-   * Style override for the input element when comboBox is disabled.
+   * Style override for the input element when ComboBox is disabled.
    */
   inputDisabled: IStyle;
 
   /**
-   * Styles for the error Message text of the comboBox.
+   * Styles for the error Message text of the ComboBox.
    */
   errorMessage: IStyle;
 
@@ -362,7 +364,7 @@ export interface IComboBoxStyles {
   optionsContainerWrapper: IStyle;
 
   /**
-   * Styles for the container of all the Combobox options
+   * Styles for the container of all the ComboBox options.
    * Includes the headers and dividers.
    */
   optionsContainer: IStyle;
@@ -388,41 +390,15 @@ export interface IComboBoxStyles {
  */
 export interface IComboBoxOptionStyles extends IButtonStyles {
   /**
-   * Styles for the text inside the comboBox option.
+   * Styles for the text inside the ComboBox option.
    * This should be used instead of the description
    * inside IButtonStyles because we custom render the text
-   * in the comboBox options.
+   * in the ComboBox options.
    */
   optionText: IStyle;
 
   /**
-   * Styles for the comboBox option text's wrapper.
+   * Styles for the ComboBox option text's wrapper.
    */
   optionTextWrapper: IStyle;
-}
-
-/**
- * {@docCategory ComboBox}
- */
-export interface IComboBoxClassNames {
-  container: string;
-  label: string;
-  root: string;
-  input: string;
-  errorMessage: string;
-  callout: string;
-  optionsContainer: string;
-  header: string;
-  divider: string;
-  optionsContainerWrapper: string;
-  screenReaderText: string;
-}
-
-/**
- * {@docCategory ComboBox}
- */
-export interface IComboBoxOptionClassNames {
-  optionText: string;
-  root: string;
-  optionTextWrapper: string;
 }
