+++ packages/react/src/components/ComboBox/index.ts
@@ -1,4 +1,3 @@
-export { SelectableOptionMenuItemType } from '@fluentui/react-internal/lib/SelectableOption';
 export * from './ComboBox';
 export * from './ComboBox.types';
 export * from './VirtualizedComboBox';
